==> src/r/scrna/07_scrna_twodonor_annotate.Rmd <==

---
title: "Two Donor scRNAseq"
output:
  html_document:
    df_print: paged
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = here::here('..','html'))})
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(hdf5r)
library(SeuratDisk)
library(dplyr)
library(tidyr)
library(tidyr)
library(Seurat)
library(patchwork)
library(cluster)
library(parallelDist)
library(ggplot2)
library(data.table)
library(cowplot)
library(future)
library(BiocParallel)
library(ggrepel)
library(scales)
library(MBA)
library(dendsort)
library(ggrastr)
library(scales)

reload_data=F
source(here::here('r','fig_colors.R'))
source(here::here('r','scrna','scrna_analysis_fxns.R'))

options(future.globals.maxSize = 14000 * 1024^2)
plan("sequential")
plan()

reload_h5_data <- T
rerun_geneset_scores <- F

car_colors_umap <- car_colors
names(car_colors_umap)[1] <- '4-1BB'
names(car_colors_umap)[8] <- 'Zeta'
names(car_colors_umap)[7] <- 'CD30'
car_colors_umap <- c(car_colors_umap, 'Untransduced'='grey60')
umap_car_order <- c("CD28","4-1BB","BAFF-R","TACI","CD40","CD30","Zeta","KLRG1","Untransduced")
car_colors_umap <- car_colors_umap[umap_car_order]

```

## Load annotated subtypes

```{r}

plots <- list()

if (reload_h5_data) {

  #CD8
  cd8_all.h5.file <- here::here(
    '..','..','scrnaseq_tcsl154',
    'rds','scrna.wnn.int.cd8.all.annotated.h5Seurat')
  scrna_cd8_all <- LoadH5Seurat(cd8_all.h5.file, overwrite=T)
  
  #CD4
  cd4_all.h5.file <- here::here(
    '..','..','scrnaseq_tcsl154',
    'rds','scrna.wnn.int.cd4.all.annotated.h5Seurat')
  scrna_cd4_all <- LoadH5Seurat(cd4_all.h5.file, overwrite=T)
  
  #All cells
  all.h5.file <- here::here(
    '..','..','scrnaseq_tcsl154',
    'rds','scrna.wnn.int.all.annotated.h5Seurat')
  scrna_all <- LoadH5Seurat(all.h5.file, overwrite=T)
}

# new annotations

# STIM-GLYC - Glycolytic
# STIM-OXPHOS - OXPHOS
# STIM-IFNG - IFNG
# STIM-INACT - Cytokine-low
# MEM-CD29 - Memory
# Naive-CD62L - Naive-CD62L
# Naive-CD7 - Naive-CD7
# TH2 - Th2
# STIM-UNTR - Baseline

# mapping from original to new cluster names
cluster_name_map <- list(
  'STIM-GLYC'='Glycolytic',
  'STIM-OXPHOS'='OXPHOS',
  'STIM-IFNG'='IFNG',
  'STIM-GNLY'='Granulysin',
  'STIM-INACT'='Cytokine-low',
  'STAT1-IRF1'='STAT1-IRF1',
  'MEM-CD29'='Memory',
  'NAIVE-CD62L'='Naive/CD62L',
  'NAIVE-CD7'='Naive/CD7',
  'TH2'='Th2',
  'STIM-UNTR'='Baseline'
)

# make dt mapping old names, new names, colors
cluster_name_dt <- data.table(
  Subtype=names(cluster_name_map), subtype_lbl=unlist(cluster_name_map))
cluster_name_dt <- data.table(
  Subtype=names(all_colors), color=unlist(all_colors))[
  cluster_name_dt, on='Subtype']

cd4_order <- c(
  'STIM-GLYC',
  'STIM-OXPHOS',
  'STIM-IFNG',
  'STAT1-IRF1',
  'TH2',
  'STIM-UNTR',
  'MEM-CD29',
  'NAIVE-CD62L'
)

cd8_order <- c(
  'STIM-GLYC',
  'STIM-OXPHOS',
  'STIM-IFNG',
  'STIM-GNLY',
  'STIM-INACT',
  'STIM-UNTR',
  'MEM-CD29',
  'NAIVE-CD7',
  'NAIVE-CD62L'
)

comb_order <- c(
  "STIM-GLYC",
  "STIM-OXPHOS",
  "STIM-IFNG",
  "STIM-GNLY",
  "STIM-INACT",
  "STIM-UNTR",
  "STAT1-IRF1",
  "TH2",
  "MEM-CD29",
  "NAIVE-CD7",
  "NAIVE-CD62L")

```

## UMAP Plots

```{r}

umap_dt <- data.table(cbind(scrna_all@meta.data, scrna_all@reductions$wnn.umap@cell.embeddings))

# add cluster labels
umap_dt <- umap_dt[, subtype_lbl := NULL][cluster_name_dt, on='Subtype']
umap_dt[, subtype_lbl := paste(t_type, subtype_lbl, sep='-')]

# invert x axis so cd4 is on left (comes first)
umap_dt[, wnnUMAP_1 := -wnnUMAP_1]

umap_dt <- umap_dt[Subtype != 'REMOVE' & !is.na(Subtype) & car != 'K_only' & Subtype != 'NA']

umap_car_order <- c("CD28","4-1BB","BAFF-R","TACI","CD40","CD30","Zeta","KLRG1","Untransduced")

cluster_name <- 'Subtype'

labels_dt <- umap_dt[, 
  list(wnnUMAP_1=mean(wnnUMAP_1), wnnUMAP_2=mean(wnnUMAP_2)),
  by=c('subtype_lbl','t_type','Subtype')]

plots$subtypes <- ggplot(umap_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, color=Subtype)) +
  rasterise(geom_point(size=0.1, alpha=0.25), dpi=500) + 
  coord_fixed() +
  geom_label_repel(
    data=labels_dt,
    aes(x=wnnUMAP_1, y=wnnUMAP_2, label=subtype_lbl),
    fill= alpha(c("white"),0.75), force=0.5, size=3) + 
  scale_color_manual(values=all_colors) +
  theme_minimal() + 
  theme(
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    panel.grid=element_blank())

inset_umap_point_size = 0.05

#per phase
plots$per_phase <- ggplot(
    umap_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, color=Phase)) +
  rasterise(geom_point(size=inset_umap_point_size*0.01, alpha=0.05), dpi=500) + 
  coord_fixed() +
  scale_color_manual(values=phase_colors) +
  theme_minimal() + 
  guides(color = guide_legend(override.aes = list(size = 5, alpha=1))) +
  theme(
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())

#per donor
plots$per_donor <- ggplot(
    umap_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, 
      color=factor(donor, labels=c('A','B')))) +
  rasterise(geom_point(size=inset_umap_point_size*0.1, alpha=0.25), dpi=500) + 
  scale_color_discrete('Donor') +
  coord_fixed() +
  theme_minimal() + 
  guides(color = guide_legend(override.aes = list(size = 5, alpha=1))) +
  theme(
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())

plots$per_donor_sep <- ggplot(
    umap_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, 
      color=factor(donor, labels=c('A','B')))) +
  rasterise(geom_point(
    data=copy(umap_dt)[, 'donor' := NULL],
    size=inset_umap_point_size*0.01, color='grey90'), dpi=500) +
  rasterise(geom_point(size=inset_umap_point_size*0.01, alpha=0.05), dpi=500) + 
  scale_color_discrete('Donor') +
  coord_fixed() +
  facet_wrap(~donor) +
  theme_minimal() + 
  guides(color = guide_legend(override.aes = list(size = 5, alpha=1))) +
  theme(
    strip.text=element_blank(),
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())

#stimulation
cluster_stim <- unique(data.table(umap_dt)[, list(n_clust_car= .N), 
  by=c(cluster_name, "car", "k_type","t_type")][, 
    pct_clust_car := n_clust_car/sum(n_clust_car), 
    by=.(car, k_type, t_type)][,
      cd19_clust_car_l2fc := log2(
          pct_clust_car[k_type == ifelse(.BY[2] == 'Untransduced', 'bead_stim', 'cd19+')]/
            pct_clust_car[k_type == 'baseline']), 
      by=.(get(cluster_name), car, t_type)][, 
        cd19_clust_act_l2fc := mean(
            cd19_clust_car_l2fc[!(car %in% c('KLRG1','Untransduced')) | k_type == 'bead_stim'], na.rm=T), 
        by=get(cluster_name)][, list(Subtype, t_type, cd19_clust_act_l2fc)])

cluster_stim[is.nan(cd19_clust_act_l2fc), cd19_clust_act_l2fc := 100]

plots$activation <- ggplot(
    umap_dt[cluster_stim, on=c('t_type','Subtype')][sample(.N)], 
    aes(x=wnnUMAP_1, y=wnnUMAP_2, 
      color=cd19_clust_act_l2fc)) +
  rasterise(geom_point(size=inset_umap_point_size*0.1, alpha=0.25), dpi=500) + 
  scale_color_distiller(
    'Stim:Unstim',
    palette='PuOr', direction=-1, oob=scales::squish, limits=c(-5,5),
    breaks=c(-5,-3,0,3,5),
    labels=rev(c('32:1','8:1','1:1','1:8','1:32'))) +
  coord_fixed() +
  theme_minimal() + 
  theme(
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    panel.grid=element_blank(),
    legend.justification = c(0,1),
    legend.key.width=unit(0.4, 'lines'),
    legend.key.height=unit(0.8, 'lines'))

#activation v2
plots$activation2 <- ggplot(
    umap_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, 
      color=factor(k_type, levels=c('baseline','cd19+','bead_stim'), labels=c('Unstim','CD19 Stim', 'Untr\n+CD3/CD28 Beads')))) +
  rasterise(geom_point(
    data=copy(umap_dt)[, 'donor' := NULL],
    size=inset_umap_point_size*0.01, color='grey90'), dpi=500) +
  rasterise(geom_point(size=inset_umap_point_size*0.01, alpha=0.2), dpi=500) + 
  scale_color_manual('Stimulation',
    values=c(
      brewer.pal(11,'PuOr')[c(9,3)],
      brewer.pal(11,'PiYG')[c(9)]),
      ) +
  coord_fixed() +
  theme_minimal() + 
  guides(color = guide_legend(
    override.aes = list(size = 5, alpha=1))) +
  theme(
    strip.text=element_blank(),
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())

combined_inset_umap <- plot_grid(
  plots$per_phase,
  NULL,
  plots$per_donor,
  NULL,
  plots$activation2,
  ncol=1,
  rel_heights=c(1,-0.1,1,-0.1,1))

# ratio = 1.8
plots$combined_umap <- (plots$subtypes + theme(legend.position='none') | 
    (plots$per_phase / plots$per_donor / plots$activation)) + plot_layout(widths=c(4.5,1))

plots$combined_umap_vert <- ((plots$subtypes + theme(legend.position='none')) / 
    (plots$per_phase | plots$per_donor | plots$activation)) + plot_layout(heights=c(4.5,1))

plots$combined_umap_vert_v2 <- ((plots$subtypes + theme(legend.position='none')) / 
  (plots$per_phase | plots$per_donor_sep | plots$activation) + plot_layout(widths=c(1,2,1))
) + plot_layout(heights=c(4.5,1))


```

## Individual CAR UMAPs

```{r}

car_umap_grid_dt <- umap_dt[car != 'CD30']
bead_stim_grid_dt <- car_umap_grid_dt[
  car == 'Untransduced' & k_type %in% c('bead_stim','baseline')][,
    car := 'CD3/CD28 Beads'][
      k_type == 'bead_stim', k_type := 'cd19+']
car_umap_grid_dt <- rbind(
  car_umap_grid_dt[!(car == 'Untransduced' & k_type %in% c('bead_stim'))],
  bead_stim_grid_dt)

car_umap_grid_dt[, k_type := factor(k_type, labels=c('Unstimulated','Stimulated'))]
car_umap_grid_dt[, car := factor(car, levels=c(umap_car_order, 'CD3/CD28 Beads'))]

# sample evenly from each CAR and donor
sampling_dt <- car_umap_grid_dt[, .N, by=c('k_type', 'car','donor')][,
  list(least_car=car[which.min(N)], least_N=min(N)), by=c('k_type','donor')]

sampled_umap <- car_umap_grid_dt[
  sampling_dt, on=c('donor','k_type')][,
    .SD[sample(.N, least_N),], by=c('car','donor','k_type')]

plots$per_car_umap <- ggplot(sampled_umap, aes(x=wnnUMAP_1, y=wnnUMAP_2, color=
    factor(Subtype,
      levels=cluster_name_dt[comb_order, on='Subtype']$Subtype,
      labels=cluster_name_dt[comb_order, on='Subtype']$subtype_lbl))) +
  rasterise(geom_point(
    data=copy(car_umap_grid_dt)[, c('car','k_type') := list(NULL,NULL)],
    size=0.1, color='grey90'), dpi=200) + 
  rasterise(geom_point(size=0.05, alpha=0.1), dpi=200) +
  coord_fixed() +
  scale_color_manual('Cluster', values= cluster_name_dt[, setNames(color, subtype_lbl)]) +
  facet_grid(k_type~car, switch = "y") +
  guides(color = guide_legend(override.aes = list(size = 5, alpha=1), nrow=1)) +
  theme_minimal() + 
  theme(
    legend.position="bottom",
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    panel.border=element_rect(color='grey50', fill=NA),
    panel.grid=element_blank())
```

## Paper Correlation UMAPs

```{r}

### Paper gene modules

DefaultAssay(scrna_all) <- 'SCT_INT'

Sheih_et_al_IRF <- c('CCL4','CCL5','CD27','CD74','FGFBP2','GNLY','GZMA','GZMB',
  'GZMH','GZMK','HLA-DRA','HLA-DRB1','HLA-DRB5','IFNG','LAG3', 'NKG7','STMN1')

Nicolet_et_al_CD29core <- c('PRSS23', 'FLNA', 'SELPLG', 'YWHAQ', 'ANXA1',
  'CX3CR1, PXN', 'FGFBP2', 'GZMH', 'TOB1', 'CD52', 'LITAF', 'FCGR3A',
  'ITGB1', 'LGALS1', 'HOPX', 'GZMB', 'GNLY', 'PATL2')

if (!('Nicolet_et_al_CD29core1' %in% names(scrna_all@meta.data))) 
  scrna_all <- AddModuleScore(scrna_all,
    features=list('1'=Nicolet_et_al_CD29core),name='Nicolet_et_al_CD29core')
if (!('Sheih_et_al_IRF1' %in% names(scrna_all@meta.data))) 
  scrna_all <- AddModuleScore(scrna_all,
    features=list('1'=Sheih_et_al_IRF),name='Sheih_et_al_IRF')

umap_dt <- data.table(cbind(scrna_all@meta.data, scrna_all@reductions$wnn.umap@cell.embeddings))
# invert x axis so cd4 is on left (comes first)
umap_dt[, wnnUMAP_1 := -wnnUMAP_1]

#reorder cars
umap_dt[, car := factor(car, levels=umap_car_order)]

umap_dt <- umap_dt[Subtype != 'REMOVE' & !is.na(Subtype) & car != 'K_only' & Subtype != 'NA']

umap_car_order <- c("CD28","4-1BB","BAFF-R","TACI","CD40","CD30","Zeta","KLRG1","Untransduced")

interpolate_score <- function(umap_data, resolution, score_name, plot=T, pal_name) {

  data <- umap_data[,list(t_type, wnnUMAP_1, wnnUMAP_2, score=get(score_name))]
  data[, scaled := score]
  data[scaled < 0, scaled := 0]
  data[, scaled := scales::rescale(scaled, to=c(0,1)), by=c('t_type')]
  
  # remove single point outliers (mostly in dead zone between cd4/cd8)
  data_no_outliers <- data[!(-wnnUMAP_1 > 1 & -wnnUMAP_1 < 3 & wnnUMAP_2 > -1.5)]
  
  grid_resolution=500
  
  # perform surface approximation from bivariate scattered data using B-splines
  mba <- mba.surf(data_no_outliers[, 
    list(wnnUMAP_1, wnnUMAP_2, scaled)], grid_resolution, grid_resolution)
  mba_interp <- data.table(
    as.vector(rep(mba$xyz.est$x, grid_resolution)),
    as.vector(rep(mba$xyz.est$y, each=grid_resolution)),
    as.vector(mba$xyz.est$z))
  names(mba_interp) <- c('wnnUMAP_1','wnnUMAP_2','score')
  mba_interp[is.na(score), score := min(mba_interp$score, na.rm=T)]
  
  if(plot) {
    return(ggplot(data, aes(x=wnnUMAP_1, y=wnnUMAP_2)) +
      rasterise(geom_point(size=0.001, color='grey50', alpha=0.01), dpi=200) +
      rasterise(geom_contour_filled(data=mba_interp, aes(z=score), alpha=0.8, bins=9)) +
      scale_fill_manual(values=c('white','white',brewer.pal(9,pal_name)[3:9])) + 
      coord_fixed() +
      theme_minimal() +
      theme(
        axis.text=element_blank(),
        axis.line = element_blank(),
        axis.title=element_blank(),
        panel.border=element_rect(color='grey50', fill=NA),
        panel.grid=element_blank()))
  } else {
    return(mba_interp)
  }
}


plots$paper_correlation_umap <- (
  interpolate_score(umap_dt, 500, 'Sheih_et_al_IRF1', pal_name='Greys') + 
    ggtitle('Sheih et al. 2020', 
      subtitle='Clonotypes enriched after\nengraftment in CAR patients') | 
  (interpolate_score(umap_dt, 500, 'Nicolet_et_al_CD29core1', pal_name='Greys') +
    ggtitle('Nicolet et al. 2020', 
      subtitle='CD29+ Cytotoxic signature,\npredicts survival in Melanoma'))) &
  theme(
    legend.position='none',
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))

umap_dt[, t_subtype := interaction(t_type, Subtype)]

umap_dt[, t_subtype := factor(t_subtype, 
  levels=c(paste('CD4',cd4_order,sep='.'), paste('CD8',cd8_order,sep='.')))]

std <- function(x) sd(x)/sqrt(length(x))

# get subtype significance data
get_sig_data <- function(subtype_dt, col_name) {
  subtype_lm <- with(subtype_dt, lm(formula = score ~ 0 + get(col_name)))
  ci_data <- data.table(
    confint(subtype_lm, level=0.99), keep.rownames=T)[,
      rn := gsub('get\\(col_name\\)','', rn)]
  padj <- summary(subtype_lm)$coefficients[,4]
  padj <- data.table(rn= gsub('get\\(col_name\\)','', names(padj)), padj=padj)
  sig_data <- merge(padj, ci_data, on='rn')
  setnames(sig_data, 'rn', col_name)
  names(sig_data)[c(3,4)] <- c('lwr','upr')
  return(sig_data)
}

make_minidot_plot <- function(umap_data, module_name, rescale_t_types=F, reorder_x=F) {
  
  if (rescale_t_types) {
    umap_data <- umap_data[, score := scale(get(module_name)), by=c('t_type')]
  } else {
    umap_data <- umap_data[, score := scale(get(module_name))]
  }
  
  umap_means <- umap_data[, 
    list(Subtype=Subtype[1], t_type=t_type[1], score=mean(score), stdev=std(score)), 
      by=c('t_subtype','donor')]
  
  sig_data <- umap_data[, get_sig_data(.SD, 'Subtype'), by=c('donor','t_type')]
  umap_means <- merge(umap_means, sig_data, on=c('t_type','donor','Subtype'))
  
  umap_means[padj < 0.01, sig_lbl := '*']
  umap_means[padj < 0.001, sig_lbl := '**']
  umap_means[padj < 0.0001, sig_lbl := '***']
  
  if (reorder_x) {
    type_ranks <- umap_means[, list(mean(-score)), by='t_subtype'][, t_subtype[order(V1)]]
    umap_means[, t_subtype := factor(t_subtype, levels=type_ranks)]
  }
  
  plot <- ggplot(umap_means, aes(
      x=t_subtype, y=score, fill=Subtype, 
      color=Subtype, group=interaction(donor,t_subtype))) +
    geom_linerange(
      aes(ymin=lwr, ymax=upr), size=0.5,
      color='grey30',
      position=position_dodge(1)) +
    geom_point(
       size=2,
       position=position_dodge(1)) + 
    # geom_text(
    #   size=5,
    #   aes(label=sig_lbl),
    #   vjust=-1,
    #   position=position_dodge(1)) +
    facet_wrap(~t_type, scales='free_x') +
    scale_fill_manual(values=all_colors) +
    scale_color_manual(values=all_colors) +
    geom_vline(xintercept = -Inf) + geom_hline(yintercept= -Inf) +
    geom_hline(yintercept = 0, linetype=2, color='grey80') + 
    scale_x_discrete('', labels=setNames(
      gsub('CD\\d\\.','',levels(umap_dt$t_subtype)), levels(umap_dt$t_subtype))) +
    theme_minimal() +
    theme(
      legend.position = 'none',
      axis.title.y = element_blank(),
      axis.text.x = element_blank(),
      panel.background=element_rect(fill=NA, color=NA),
      panel.border=element_rect(fill=NA, color=NA),
      strip.text = element_text(vjust = -1.0), 
      panel.spacing.y = unit(-2, "lines"),
      panel.grid = element_blank())
  
  return(plot)
}

make_minidot_plot_car <- function(umap_data, module_name, run_test=F, rescale_t_types=F, reorder_x=F) {
  
  if (rescale_t_types) {
    umap_data <- umap_data[, score := scale(get(module_name)), by=c('t_type')]
  } else {
    umap_data <- umap_data[, score := scale(get(module_name))]
  }
  
  # plot means
  umap_means <- umap_data[, 
    list(score=mean(score), stdev=std(score)), 
      by=c('t_type','donor','car')]
  
  #significance
  sig_data <- umap_data[, get_sig_data(.SD,'car'), by=c('donor','t_type')]
  umap_means <- merge(umap_means, sig_data, on=c('t_type','donor','car'))
  
  umap_means <- umap_means[, t_car := interaction(t_type, car)]
  
  if (reorder_x) {
    car_ranks <- umap_means[, list(mean(-score)), by=c('t_car')][, t_car[order(V1)]]
    umap_means[, t_car := factor(t_car, levels=car_ranks)]
  }
  
  plot <- ggplot(umap_means, aes(
      x=t_car, y=score, fill=car, 
      color=car, group=interaction(donor,car))) +
    geom_linerange(
      aes(ymin=lwr, ymax=upr), size=0.5,
      color='grey30',
      position=position_dodge(1)) +
     geom_point(
       size=2,
       position=position_dodge(1)) + 
    facet_wrap(~t_type, scales='free_x') +
    facet_wrap(~t_type, scales='free_x') +
    scale_fill_manual(values=car_colors_umap) +
    scale_color_manual(values=car_colors_umap) +
    geom_vline(xintercept = -Inf) + geom_hline(yintercept= -Inf) +
    geom_hline(yintercept = 0, linetype=2, color='grey80') + 
    theme_minimal() +
    theme(
      legend.position = 'none',
      axis.title.y = element_blank(),
      axis.title.x = element_blank(),
      axis.text.x = element_blank(),
      panel.background=element_rect(fill=NA, color=NA),
      panel.border=element_rect(fill=NA, color=NA),
      strip.text = element_text(vjust = -1.0), 
      panel.spacing.y = unit(-2, "lines"),
      panel.grid = element_blank())
  
  return(plot)
}

umap_nocd30_dt <- umap_dt[car != 'CD30']

cluster_zscore_plots <- (((make_minidot_plot(umap_dt, 'Sheih_et_al_IRF1') + labs(y='cluster z-score') + theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5))) | make_minidot_plot(umap_dt, 'Nicolet_et_al_CD29core1')))

car_zscore_plots <- (((make_minidot_plot_car(umap_nocd30_dt, 'Sheih_et_al_IRF1') + labs(y='CAR z-score') + theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5))) | make_minidot_plot_car(umap_nocd30_dt, 'Nicolet_et_al_CD29core1')))

plots$combined_paper_umap <- (
  (plots$paper_correlation_umap & theme(plot.margin= unit(c(0, 0, 0, 0), "points"))) /
    cluster_zscore_plots / car_zscore_plots) +
      plot_layout(heights=c(4,1,1))

# 
# # kw/ dunn tests
# kw_tests <- umap_dt[, dunn_test(formula = Sheih_et_al_IRF1 ~ Subtype, data=.SD), by=c('t_type','donor')]
# ggplot(kw_tests) + geom_tile(aes(x=group1, y=group2, fill=-log(p.adj), color=p.adj < 0.05)) + facet_wrap(~t_type+donor, scales='free') + scale_fill_distiller(palette='Greens', oob=scales::oob_squish_any, limits=c(-log(0.05), -log(0.001)), direction=1) + theme(axis.text.x=element_text(angle = 90, vjust = 1, hjust=0.5))
# 
# kw_tests <- umap_dt[, dunn_test(formula = Nicolet_et_al_CD29core1 ~ Subtype, data=.SD), by=c('t_type','donor')]
# ggplot(kw_tests) + geom_tile(aes(x=group1, y=group2, fill=-log(p.adj), color=p.adj < 0.05)) + facet_wrap(~t_type+donor, scales='free') + scale_fill_distiller(palette='Greens', oob=scales::oob_squish_any, limits=c(-log(0.05), -log(0.001)), direction=1) + theme(axis.text.x=element_text(angle = 90, vjust = 1, hjust=0.5))
# 

plots$combined_paper_umap

```

# CAR/Cluster Dotplots

```{r}
car_pct_dotplot <- function(
    car_metadata, subtype_order, max_size=15, flip=T, legend.pos='right', car_text_relsize=1.2) {
  
  car_clust_pct_act <- data.table(car_metadata)[, .N, 
    by=c('car','Subtype')][, 
      pct_car := N/sum(N), by='car'][,
        clust_count := sum(N), by='Subtype'][,
          log2_clust_enrich := log2(pct_car/mean(pct_car)), by='Subtype']
  
  car_clust_pct_act[, log2_ce_trunc := ifelse(
    test= abs(log2_clust_enrich) > 1, 
    yes= 1 * sign(log2_clust_enrich), 
    no= log2_clust_enrich)]
  
  pct_dotplot <- ggplot(car_clust_pct_act) + 
    geom_point(aes(
        x= factor(Subtype,
          levels=cluster_name_dt[rev(subtype_order), on='Subtype']$Subtype,
          labels=cluster_name_dt[rev(subtype_order), on='Subtype']$subtype_lbl),
        y=factor(car, umap_car_order, gsub('Untransduced','Untrans.', umap_car_order)),
        fill=log2_ce_trunc, size=pct_car),
      pch=21, color='grey30') + 
    scale_fill_distiller('log2\nCAR Enrichment', palette='PRGn', direction=1, 
      breaks=c(-1,-0.5,0,0.5,1), labels=c('< -1','-0.5','0','0.5','> 1')) + 
    scale_size('% CAR Cells', range=c(1,max_size), 
      breaks=c(0.1, 0.25, 0.4), labels=c('10%','25%','40%')) + 
    theme_minimal() + 
    theme(
      legend.key.width=unit(0.4, 'lines'),
      axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + 
    labs(y='', x='')
  
  if (flip)
    pct_dotplot <- pct_dotplot + coord_flip()
  
  unique_car_v_subsets <- unique(car_clust_pct_act[, list(car, Subtype)])
  n_lines <- unique_car_v_subsets[, length(unique(Subtype))]
  
  if (flip) {
    cluster_idents <- ggplot(unique_car_v_subsets) + 
      geom_tile(aes(
        y=factor(Subtype,
          levels=cluster_name_dt[rev(subtype_order), on='Subtype']$Subtype,
          labels=cluster_name_dt[rev(subtype_order), on='Subtype']$subtype_lbl),
        fill=factor(Subtype,
          levels=cluster_name_dt[rev(subtype_order), on='Subtype']$Subtype,
          labels=cluster_name_dt[rev(subtype_order), on='Subtype']$subtype_lbl),
        x='a'), width=0.9, height=0.9) + 
      #geom_hline(color='white', yintercept=0.5+c(1:n_lines-1), size=2) +
      scale_x_discrete(expand=c(0,0)) + 
      scale_y_discrete(expand=c(0,0)) +
      scale_fill_manual(values=cluster_name_dt[, setNames(color, subtype_lbl)]) +
      theme_minimal() +
      coord_fixed(ratio=3) +
      theme(
        legend.position = 'none',
        axis.title.x= element_blank(),
        axis.title.y=element_blank(),
        axis.text.x=element_blank(),
        panel.background = element_blank(),
        panel.grid = element_blank(),
        plot.margin = unit(c(0, 22, 5.5, 5.5), "points"))
  
    pct_dotplot_combined <- cluster_idents + theme(plot.margin=unit(c(5.5,5.5,5.5,-5.5), "points")) | 
      (pct_dotplot + theme(
        legend.position=legend.pos,
        axis.text.y=element_blank(),
        axis.text.x=
          element_text(color=car_colors_umap[names(car_colors_umap) != 'CD30'],
          face='bold',
          size=rel(car_text_relsize)),
        strip.text.x=element_blank(),
        axis.title.y=element_blank(),
        plot.margin = unit(c(5.5, -5.5, 5.5, 5.5), "points")))
  } else {
    cluster_idents <- ggplot(unique_car_v_subsets) + 
      geom_tile(aes(
        x=factor(Subtype,
          levels=cluster_name_dt[rev(subtype_order), on='Subtype']$Subtype,
          labels=cluster_name_dt[rev(subtype_order), on='Subtype']$subtype_lbl),
        fill=factor(Subtype,
          levels=cluster_name_dt[rev(subtype_order), on='Subtype']$Subtype,
          labels=cluster_name_dt[rev(subtype_order), on='Subtype']$subtype_lbl),
        y='a'),
        width=0.9, height=0.9) + 
      #geom_hline(color='white', yintercept=0.5+c(1:n_lines-1), size=2) +
      scale_x_discrete(expand=c(0,0)) + 
      scale_y_discrete(expand=c(0,0)) +
      scale_fill_manual(values=cluster_name_dt[, setNames(color, subtype_lbl)]) +
      theme_minimal() +
      coord_fixed(ratio=1/3) +
      theme(
        legend.position = 'none',
        axis.title.y= element_blank(),
        axis.title.x=element_blank(),
        axis.text.y=element_blank(),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.background = element_blank(),
        panel.grid = element_blank(),
        plot.margin = unit(c(0, 5.5, 22, 5.5), "points"))
    
    pct_dotplot_combined <- (
      pct_dotplot + theme(
        legend.position=legend.pos,
        axis.text.x=element_blank(),
        strip.text.x=element_blank(),
        # color CAR names
        axis.text.y=
          element_text(color=car_colors_umap[names(car_colors_umap) != 'CD30'],
            face='bold',
            size=rel(car_text_relsize)),
        axis.title.y=element_blank(),
        plot.margin = unit(c(5.5, 5.5, -5.5, 5.5), "points"))) /
      cluster_idents + theme(plot.margin=unit(c(-5.5,5.5,5.5,5.5), "points"))
  }
  
  return(pct_dotplot_combined)
}

plots$cd8_dotplot <- car_pct_dotplot(
  data.table(scrna_cd8_all@meta.data)[
    k_type == 'cd19+' & !is.na(Subtype) &
    car != 'CD30' & Subtype != 'REMOVE' &
    Subtype != 'NA'], cd8_order, max_size=9)

plots$cd4_dotplot <- car_pct_dotplot(
  data.table(scrna_cd4_all@meta.data)[
    k_type == 'cd19+' & !is.na(Subtype) &
    car != 'CD30' & Subtype != 'REMOVE' &
    Subtype != 'NA'], cd4_order, max_size=9)

plots$cd8_unstim <- car_pct_dotplot(
  data.table(scrna_cd8_all@meta.data)[
    k_type == 'baseline' & !is.na(Subtype) & car != 'K_only' & !grepl('STIM',Subtype) &
    car != 'CD30' & Subtype != 'REMOVE' &
    Subtype != 'NA'
      ], cd8_order, max_size=9)

plots$cd4_unstim <- car_pct_dotplot(
  data.table(scrna_cd4_all@meta.data)[
    k_type == 'baseline' & !is.na(Subtype) & car != 'K_only' & !grepl('STIM',Subtype) &
    car != 'CD30' & Subtype != 'REMOVE' &
    Subtype != 'NA'], cd4_order, max_size=9)

plots$car_dotplot <- (
  ((plots$cd4_dotplot + ggtitle('CD4 Stimulated') + 
      theme(legend.position='none', axis.text.x=element_blank())) / 
     (plots$cd8_dotplot + ggtitle('CD8 Stimulated') +
        # color CAR names
        theme(
          axis.text.x=
            element_text(color=car_colors_umap[names(car_colors_umap) != 'CD30'],
          face='bold',
          size=rel(1))))) & 
      theme(plot.title.position="plot")
  ) + 
  plot_layout(guides='collect', heights=c(8,9)) & 
  theme(plot.margin=unit(c(0,0,0,0),'points'))

plots$car_dotplot_unstim <- (
  ((plots$cd4_unstim + ggtitle('CD4 Baseline') + 
      theme(legend.position='none', axis.text.x=element_blank())) / 
     (plots$cd8_unstim + ggtitle('CD8 Baseline'))) & 
      theme(plot.title.position="plot")
  ) + 
  plot_layout(guides='collect', heights=c(4,3)) & 
  theme(plot.margin=unit(c(0,0,0,0),'points'))

plots$car_dotplot_unstim_vert <- (
  ((car_pct_dotplot(
      data.table(scrna_cd4_all@meta.data)[
        k_type == 'baseline' & !is.na(Subtype) & car != 'K_only' & !grepl('STIM',Subtype) &
        car != 'CD30' & Subtype != 'REMOVE' &
        Subtype != 'NA'], cd4_order, max_size=9, legend.pos='none', car_text_relsize=0.9) +
      ggtitle('CD4 Baseline') + 
      theme(legend.position='none')) |
     (car_pct_dotplot(
      data.table(scrna_cd8_all@meta.data)[
        k_type == 'baseline' & !is.na(Subtype) & car != 'K_only' & !grepl('STIM',Subtype) &
        car != 'CD30' & Subtype != 'REMOVE' &
        Subtype != 'NA'], cd8_order, max_size=9, legend.pos='none', car_text_relsize=0.9) +
      ggtitle('CD8 Baseline'))) & 
      theme(plot.title.position="plot")
  ) + 
  plot_layout(widths=c(1,10,10)) & 
  theme(plot.margin=unit(c(0,0,0,0),'points'))

plots$car_dotplot_all <- (
  (
    (plots$cd4_dotplot + ggtitle('CD4 Stimulated') + 
      theme(legend.position='none', axis.text.x=element_blank())) / 
    (plots$cd8_dotplot + ggtitle('CD8 Stimulated') +
      theme(legend.position='none', axis.text.x=element_blank())) /
    (plots$cd4_unstim + ggtitle('CD4 Baseline') + 
      theme(legend.position='none', axis.text.x=element_blank())) / 
     (plots$cd8_unstim + ggtitle('CD8 Baseline'))
  ) & theme(plot.title.position="plot")) + 
  plot_layout(guides='collect', heights=c(8,9,4,3)) & 
  theme(plot.margin=unit(c(0,0,0,0),'points'))

plots$car_dotplot_horiz <- (
  car_pct_dotplot(
    data.table(scrna_cd4_all@meta.data)[
      k_type == 'cd19+' & !is.na(Subtype) &
      car != 'CD30' & Subtype != 'REMOVE' &
      Subtype != 'NA'], cd4_order, max_size=9, flip=F) | (car_pct_dotplot(
    data.table(scrna_cd8_all@meta.data)[
      k_type == 'cd19+' & !is.na(Subtype) &
      car != 'CD30' & Subtype != 'REMOVE' &
      Subtype != 'NA'], rev(cd8_order), max_size=9, flip=F) & theme(
        legend.position='none', axis.text.y=element_blank()))) + 
  plot_layout(guides='collect')

```

## Gene Heatmaps

```{r}
find_genes_parallel <- function(seurat_obj, ident_name, assay='SCT_INT', min.pct=0.25, features=NULL, test.use='wilcox') {

  Idents(seurat_obj) <- ident_name
  idents=unique(seurat_obj[[ident_name]])[[1]]

  FindMarker.wrapper <- function(x) {

    message(paste0('Starting FindMarkers for: ',idents[x]))

    markers_dt <- data.table(
      FindMarkers(seurat_obj,
        ident.1=idents[x],
        only.pos = TRUE, min.pct=min.pct,
        features=features,
        assay=assay, test.use=test.use),
      keep.rownames=T)

    markers_dt$ident <- idents[x]

    message(paste0('Completed FindMarkers for: ',idents[x]))

    return(markers_dt)
  }
  gc()
  subset_markers <- bplapply(
    1:length(idents),
    FindMarker.wrapper,BPPARAM=MulticoreParam(12, log=T))

  subset_markers <- rbindlist(subset_markers)
  return(subset_markers)
}

# find DE genes in SCT_INT for each cluster, this takes a ton of memory/time,
# so doing it once manually in a clean environment for each subset
rerun_find_markers <- F
cd8_marker_genes.rds.file <- here::here(
     '..','..','scrnaseq_tcsl154',
     'rds',
     paste0('cd8_marker_genes_sct.rds'))
cd4_marker_genes.rds.file <- here::here(
     '..','..','scrnaseq_tcsl154',
     'rds',
     paste0('cd4_marker_genes_sct.rds'))

if (rerun_find_markers) {
  #CD8
  pruned_cd8 <- prune_cells(scrna_cd8_all)
  rm(scrna_cd8_all)
  cd8_rna_mast_markers <- find_genes_parallel(
    pruned_cd8, 'Subtype', assay='SCT_INT', min.pct=0.20, test.use='MAST')
  saveRDS(cd8_rna_mast_markers, cd8_marker_genes.rds.file)
  
  #CD4
  pruned_cd4 <- prune_cells(scrna_cd4_all)
  rm(scrna_cd4_all)
  cd4_rna_mast_markers <- find_genes_parallel(
    pruned_cd4, 'Subtype', assay='SCT_INT', min.pct=0.20, test.use='MAST')
  saveRDS(cd4_rna_mast_markers, cd4_marker_genes.rds.file)
  
} else {
  # load the already made DE gene sets
  cd4_rna_mast_markers <- readRDS(cd4_marker_genes.rds.file)
  cd8_rna_mast_markers <- readRDS(cd8_marker_genes.rds.file)
}

combined_de_markers <- rbind(
  cd4_rna_mast_markers[, t_type := 'CD4'],
  cd8_rna_mast_markers[, t_type := 'CD8'])
combined_de_markers[, t_subtype := paste(t_type, ident, sep='.')]
combined_de_markers[p_val_adj < 0.05, ident_rank := rank(-abs(avg_log2FC)), by=t_subtype]

# also tried ranking by p-value
#combined_de_markers[p_val_adj < 0.05, ident_rank := rank(-log10(p_val_adj)), by=t_subtype]

pairs <- combn(unique(combined_de_markers$t_subtype),2)

top_n <- 100

diff_genes <- rbindlist(
  apply(pairs, 2, function(i) 
    combined_de_markers[t_subtype %in% i & p_val_adj < 0.05 & ident_rank <= top_n,
      list(
        ident_1= i[1],
        ident_2= i[2],
        subtype_1=ident[t_subtype == i[1]],
        subtype_2=ident[t_subtype == i[2]],
        shared=(.N == 2 && (all(sign(avg_log2FC)) | all(sign(-avg_log2FC))))),
    by='rn'][shared == T]))

diff_gene_n <- diff_genes[
  grepl('CD4', ident_1) & grepl('CD8', ident_2)][,
    list(N=.N, subtype_1=subtype_1[1], subtype_2=subtype_2[1]), 
    by=.(ident_1, ident_2)][order(-N)]

make_paired_factor <- function(bool,flip=F) {
  if (flip)
    factor(bool, levels=c(F,T), labels=c('non-matched','matched'))
  else
    factor(bool, levels=c(T,F), labels=c('matched','non-matched'))
}

shared_subtypes <- diff_gene_n[subtype_1 != 'STIM-UNTR', intersect(
    unique(subtype_1), unique(subtype_2))]

diff_gene_n[, 
  paired_1 := make_paired_factor(subtype_1 %in% shared_subtypes, flip=F)]
diff_gene_n[, 
  paired_2 := make_paired_factor(subtype_2 %in% shared_subtypes, flip=T)]

pairs <- combn(unique(combined_de_markers$t_subtype),2)

top_n <- 100

diff_genes <- rbindlist(
  apply(pairs, 2, function(i) 
    combined_de_markers[t_subtype %in% i & p_val_adj < 0.05 & ident_rank <= top_n,
      list(
        ident_1= i[1],
        ident_2= i[2],
        subtype_1=ident[t_subtype == i[1]],
        subtype_2=ident[t_subtype == i[2]],
        shared=(.N == 2 && (all(sign(avg_log2FC)) | all(sign(-avg_log2FC))))),
    by='rn'][shared == T]))

diff_gene_n <- diff_genes[
  grepl('CD4', ident_1) & grepl('CD8', ident_2)][,
    list(N=.N, subtype_1=subtype_1[1], subtype_2=subtype_2[1]), 
    by=.(ident_1, ident_2)][order(-N)]

make_paired_factor <- function(bool,flip=F) {
  if (flip)
    factor(bool, levels=c(F,T), labels=c('non-matched','matched'))
  else
    factor(bool, levels=c(T,F), labels=c('matched','non-matched'))
}

shared_subtypes <- diff_gene_n[subtype_1 != 'STIM-UNTR', intersect(
    unique(subtype_1), unique(subtype_2))]

diff_gene_n[, 
  paired_1 := make_paired_factor(subtype_1 %in% shared_subtypes, flip=F)]
diff_gene_n[, 
  paired_2 := make_paired_factor(subtype_2 %in% shared_subtypes, flip=T)]

diff_gene_n[, subtype_1 := factor(subtype_1,
        levels=cluster_name_dt[cd4_order, on='Subtype']$Subtype,
        labels=cluster_name_dt[cd4_order, on='Subtype']$subtype_lbl)]

diff_gene_n[, subtype_2 := factor(subtype_2,
      levels=cluster_name_dt[cd8_order, on='Subtype']$Subtype,
      labels=cluster_name_dt[cd8_order, on='Subtype']$subtype_lbl)]

plots$matched_clusters <- ggplot(diff_gene_n,
    aes(
      x=subtype_1,
      y=subtype_2)) + 
  geom_tile(aes(
    fill=N/top_n)) + 
  geom_tile(aes(
    color=(as.character(subtype_1) == as.character(subtype_2) &
      subtype_1 != 'Baseline')), fill=NA, size=0.5) + 
  scale_fill_gradientn(
    '% overlap of\ntop 100\nDE genes',
    colors=c('white',brewer.pal(9,'Greens')),  label=label_percent()) + 
  theme_minimal() + 
  scale_color_manual('', guide='none', values=c(NA,'grey50')) +
  facet_grid(paired_2~paired_1, space='free', scale='free', switch='both') +
  scale_x_discrete('CD4 Clusters', guide = guide_axis(angle = 90), expand=c(0,0)) +
  scale_y_discrete('CD8 Clusters', expand=c(0,0)) + 
  theme(
    panel.grid=element_blank(),
    strip.placement = "outside",
    panel.border = element_rect(fill=NA, color='grey50'))

```

```{r}

cd4_genes <- list(
  'STIM-IFNG'= c(
    "IFNG","GZMA", "PRF1", "HLA-DRA", 
    "CD74", "CD2", "ITGB2", "EOMES", "GNLY", "LAG3"),
  'STIM-GLYC'= c(
    "IL2RA", "PGK1", "ARG2","TNFRSF18",
    "HILPDA", "BNIP3", "CTLA4", "TNFRSF9", "GAPDH"),
  'STIM-OXPHOS'= c(
    "MT-CO3", "MT-CO2", "SRM", "C1QBP",
    "HSP90AA1", "FABP5", "HSPE1", 
    "MKI67", "MT-ND1", "MT-ATP6"),
  'STAT1-IRF1'= c(
    "STAT1", "IRF1","CALM2", "ID3", "IFITM1", "IFITM2", "IFITM3","STMN1"),
  'NAIVE-CD62L'= c(
    "SELL", "LRRN3", "SOCS2", "FAM13A"),
  'STIM-UNTR'= c(
    "CCR7", "CD7", "MAL", "CD27", "CD69"),
  'TH2'= c(
    "IL5", "IL13", "GATA3", "IL4"),
  'MEM-CD29'= c(
    "S100A4", "KLRB1", "ITGB1", "CXCR3",
    "CCR5","ITGA4", "HOPX","NKG7", "NEAT1", "KLRG1"))

cd8_genes <-list(
  "NAIVE-CD7"= c(
    "CD7","TCF7","LEF1","CD27",
     "IFITM1", "IFITM2", "VIM", "STMN1", "HMGN2", "STIM1"),
  "MEM-CD29"= c(
    "S100A4", "KLRB1", "ITGB1", "CXCR3",
    "ANXA1", "KLRG1", "HOPX","ITGA1"),
  "STIM-IFNG"= c(
    "IFNG", "GZMH", "GZMB", "HLA-DRA",
    "GZMK",
    "NKG7", "CCR5", "HLA-DQA1", "CD74",
    "CD2","CCL5","CCL3",'CCL4'),
  "STIM-GLYC"= c(
    "ARG2", "HILPDA", "BNIP3", "LAYN",
    "BNIP3L", "MT2",
    "LDHA", "FAM162A", "JUNB"),
  "STIM-UNTR"= c(
    "ENO1", "PGK1", "STAT1", "CD27", "S100A10", "CD38", "IGFBP2", "NEAT1", "TNFAIP2"),
  "STIM-OXPHOS"= c(
    "C1QBP","SRM","FKBP4", "HSPD1",
    "HSP90AB1","EIF5A", "HSPH1", 
    "IL2RA"),
  "STIM-INACT"= c(
    "GZMA", "TIMP1", "TNFSF10", "PFN1"),
  "STIM-GNLY"= c(
    "GNLY", "LAYN", "TNFRSF18", "RUNX3", "CBLB", "CTSW", "LGALS3"),
  "NAIVE-CD62L"= c(
    "SELL", "IL7R", "LRRN3", "MAL", "KLF2", "TXNIP"))

make_gene_heatmap <- function(
    seurat_obj, subtype_order, max_genes=100, label_genes=NULL, 
    rescale=T, use_dendsort=T, label_width=3.25, label_font_size=2.25,
    label_line_width=22, boxed=F, return_data=F) {
  
  if (rescale)
    seurat_obj <- ScaleData(seurat_obj, assay='SCT_INT')
  
  all_genes <- AverageExpression(
    seurat_obj,
    assays = 'SCT_INT',
    features = NULL,
    return.seurat = FALSE,
    group.by = "Subtype",
    add.ident="donor",
    slot = "scale.data"
  )

  # melt, clean
  melted_genes <- reshape2::melt(all_genes[[1]])
  melted_genes <- data.table(melted_genes)
  names(melted_genes) <- c('gene','subtype','zscore')
  melted_genes[, donor := gsub('.*_(.*)','\\1', subtype)]
  melted_genes[, subtype := gsub('(.*)_(.*)','\\1', subtype)]
  melted_genes[, subtype := factor(subtype, levels=subtype_order)]
  melted_genes[, best_subtype := subtype[which.max(zscore)], by=gene]
  melted_genes[, zdiff := zscore[which.max(zscore)] - mean(zscore), by=gene]
  gene_order <- unique(melted_genes[subtype == best_subtype, gene[order(zdiff)]])
  melted_genes[, gene := factor(gene, levels=gene_order)]
  melted_genes[, max_score := max(abs(zscore)), by='gene']
  melted_genes[, mean_zscore := mean(zscore), by=c('gene','subtype')]
  
  #reorder
  new_gene_order <- c()
  for (group_i in rev(subtype_order)) {
    if (use_dendsort) {
      cast_group <- dcast(melted_genes[!is.na(subtype) & best_subtype == group_i], 
          gene ~ subtype + donor, value.var='zscore')
      
      group_gene_order <- dendsort(hclust(dist(
        cast_group[,-1])))$order
      group_genes <- as.character(cast_group$gene[group_gene_order])
    } else {
      group_genes <- melted_genes[
        !is.na(subtype) & best_subtype == group_i][
          subtype == best_subtype][order(mean_zscore), unique(as.character(gene))]
    }
    new_gene_order <- c(new_gene_order, group_genes)
  }
  
  melted_genes[, gene := factor(gene, levels=new_gene_order)]
  cast_formula <- "best_subtype ~."
  
  # subset on best genes per cluster
  best_genes <- unique(
    melted_genes[subtype == best_subtype][
      rev(new_gene_order), on='gene'][,
        list(subtype, mean_zscore), by='gene'])[order(-mean_zscore),
          gene[1:max_genes], by='subtype']$V1
  
  # best genes per cluster
  label_gene_list <- unique(unlist(label_genes, use.names=F))
  best_genes <- unique(c(best_genes, label_gene_list))
  
  #first 4 genes chosen per cluster
  glist <- list()
  for (i in 1:length(label_genes)) 
    glist[[i]] <- data.table(subtype=names(label_genes[i]), genes=label_genes[[i]][1:4])
  gene_top4 <- rbindlist(glist)
  
  # for annotation
  best_per_subtype <-unique(
    melted_genes[subtype == best_subtype][
      rev(new_gene_order), on='gene'][,
        list(subtype, mean(zscore)), by='gene'])[,
          is_best := gene %in% gene[1:max_genes], by='subtype']
  
  melted_genes[
    gene %in% best_genes &
    subtype == best_subtype &
    donor == donor[1],
      label := ifelse(gene %in% label_gene_list, as.character(gene), NA)]
    
  # alternate none with subtype name (two donors)
  row_names <- as.vector(t(cbind(rep('', length(subtype_order)), 
    cluster_name_dt[subtype_order, on='Subtype'][, subtype_lbl])))
    
  unique_donor_subset <- unique(melted_genes[, list(donor, subtype)])
  n_lines <- unique_donor_subset[, length(unique(subtype))]
  
  max_x <- length(levels(melted_genes[gene %in% best_genes, interaction(donor,subtype)]))
  
  display_genes <- melted_genes[gene %in% best_genes]
  
  separator_table <- display_genes[
      subtype == best_subtype & donor == donor[1], list(n_genes= as.double(.N)), by='subtype']
      
  separator_table[rev(subtype_order), on='subtype', pos := cumsum(n_genes)+0.5]
  separator_table[, x_pos := (n_lines-rank(pos))*2 + 0.5]
  
  label_table <- display_genes[,
    paste(na.omit(label), collapse=', '), by='subtype'][
      separator_table, on='subtype']
  
  top4_table <- gene_top4[,
    paste(na.omit(genes), collapse='\n'), by='subtype'][
      separator_table, on='subtype']
  
  de_gene_heatmap <-  de_gene_heatmap <- ggplot(display_genes) + 
    geom_tile(aes(x=interaction(donor,subtype), y=gene, fill=zscore)) +
    scale_x_discrete(expand=c(0,0)) + 
    scale_y_discrete(expand=c(0,0)) +
    coord_cartesian(clip = 'off') +
    geom_vline(color='white', xintercept=0.5+c(1:n_lines-1)*2) +
    geom_segment(data=separator_table[-.N], color='black', linetype=2, 
      aes(y=pos, yend=pos), x=0.5, xend=max_x+label_width) +
    scale_fill_distiller('Gene\nz-score',
      palette='PuOr', oob=scales::squish, limit=c(-2, 2), type='div',
      breaks=c(2,0,-2), labels=c('≥2','0','≤-2')) +
    geom_label(
      data=label_table,
      label.size=0,
      fill=NA,
      nudge_x=0.55,
      size=label_font_size,
      aes(x=max_x, y=pos-2.5, label=stringr::str_wrap(V1, label_line_width), vjust=1, hjust=0)
    ) +
    theme_minimal() +
    theme(
      axis.text.x=element_blank(),
      strip.text.y=element_blank(),
      axis.text.y=element_blank(),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(5.5, 22, 0, 5.5), "points"))
  
  de_gene_heatmap_boxed <- ggplot(display_genes) + 
    geom_tile(aes(x=interaction(donor,subtype), y=gene, fill=zscore)) +
    scale_x_discrete(expand=c(0,0)) + 
    scale_y_discrete(expand=c(0,0)) +
    geom_vline(color='white', xintercept=0.5+c(1:n_lines-1)*2) +
    geom_rect(data=separator_table,
      aes(ymax=pos-0.25, ymin=pos-n_genes+0.25, xmin=x_pos, color=subtype),
      xmax=max_x+label_width, fill=NA, size=0.6) +
    geom_rect(data=separator_table,
      aes(ymax=pos-0.25, ymin=pos-n_genes+0.25, xmin=x_pos, xmax=x_pos+2, color=subtype), fill=NA, size=0.6) +
    scale_color_manual(guide='none', values=unlist(all_colors)) +
    scale_fill_distiller('Gene\nz-score',
      palette='PuOr', oob=scales::squish, limit=c(-2, 2), type='div',
      breaks=c(2,0,-2), labels=c('≥2','0','≤-2')) +
    geom_label(
      data=top4_table,
      label.size=0,
      fill=NA,
      nudge_x=0.55,
      size=label_font_size*1.8,
      aes(x=max_x, y=pos-2.5, label=V1, vjust=1, hjust=0)) +
    coord_cartesian(clip = 'off') +
    theme_minimal() +
    theme(
      axis.text.x=element_blank(),
      strip.text.y=element_blank(),
      axis.text.y=element_blank(),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(5.5, 22, 0, 5.5), "points"))
  
  donor_idents <- ggplot(unique_donor_subset) + 
    geom_tile(aes(x=interaction(donor, subtype), fill=donor, y='a')) + 
    scale_x_discrete(labels=row_names, expand=c(0,0)) + 
    scale_y_discrete(expand=c(0,0)) +
    geom_vline(color='white', xintercept=0.5+c(1:n_lines-1)*2) +
    scale_fill_discrete('Donor', labels=c('A','B')) + 
    theme_minimal() +
    coord_fixed(ratio=1) +
    theme(
      axis.text.x=element_blank(),
      axis.text.y=element_blank(),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(0, 22, 0, 5.5), "points"))
  
  cluster_idents <- ggplot(unique_donor_subset) + 
    geom_tile(aes(x=interaction(donor, subtype), fill=subtype, y='a')) + 
    geom_vline(color='white', xintercept=0.5+c(1:n_lines-1)*2) +
    scale_x_discrete(labels=row_names, expand=c(0,0)) + 
    scale_y_discrete(expand=c(0,0)) +
    scale_fill_manual(values=all_colors) +
    theme_minimal() +
    coord_fixed(ratio=1) +
    theme(
      legend.position = 'none',
      axis.text.x = element_text(angle = 90, vjust = -0.75, hjust=1),
      axis.title.y= element_blank(),
      axis.title.x=element_blank(),
      axis.text.y=element_blank(),
      panel.background = element_blank(),
      plot.margin = unit(c(0, 22, 5.5, 5.5), "points"))
  
  if (return_data)
    melted_genes
  else if (boxed)
    return(de_gene_heatmap_boxed / donor_idents / cluster_idents)
  else
    return(de_gene_heatmap / donor_idents / cluster_idents)
}

make_gene_heatmap_horiz <- function(
    seurat_obj, subtype_order, max_genes=100, label_genes=NULL, 
    rescale=T, use_dendsort=T, label_width=5, label_font_size=2.25,
    boxed=F, return_data=F) {
  
  if (rescale)
    seurat_obj <- ScaleData(seurat_obj, assay='SCT_INT')
  
  all_genes <- AverageExpression(
    seurat_obj,
    assays = 'SCT_INT',
    features = NULL,
    return.seurat = FALSE,
    group.by = "Subtype",
    add.ident="donor",
    slot = "scale.data"
  )

  # melt, clean
  melted_genes <- reshape2::melt(all_genes[[1]])
  melted_genes <- data.table(melted_genes)
  names(melted_genes) <- c('gene','subtype','zscore')
  melted_genes[, donor := gsub('.*_(.*)','\\1', subtype)]
  melted_genes[, subtype := gsub('(.*)_(.*)','\\1', subtype)]
  melted_genes[, subtype := factor(subtype, levels=subtype_order)]
  melted_genes[, best_subtype := subtype[which.max(zscore)], by=gene]
  melted_genes[, zdiff := zscore[which.max(zscore)] - mean(zscore), by=gene]
  gene_order <- unique(melted_genes[subtype == best_subtype, gene[order(zdiff)]])
  melted_genes[, gene := factor(gene, levels=gene_order)]
  melted_genes[, max_score := max(abs(zscore)), by='gene']
  melted_genes[, mean_zscore := mean(zscore), by=c('gene','subtype')]
  
  #reorder
  new_gene_order <- c()
  for (group_i in rev(subtype_order)) {
    if (use_dendsort) {
      cast_group <- dcast(melted_genes[!is.na(subtype) & best_subtype == group_i], 
          gene ~ subtype + donor, value.var='zscore')
      
      group_gene_order <- dendsort(hclust(dist(
        cast_group[,-1])))$order
      group_genes <- as.character(cast_group$gene[group_gene_order])
    } else {
      group_genes <- melted_genes[
        !is.na(subtype) & best_subtype == group_i][
          subtype == best_subtype][order(mean_zscore), unique(as.character(gene))]
    }
    new_gene_order <- c(new_gene_order, group_genes)
  }
  
  melted_genes[, gene := factor(gene, levels=new_gene_order)]
  cast_formula <- "best_subtype ~."
  
  # subset on best genes per cluster
  best_genes <- unique(
    melted_genes[subtype == best_subtype][
      rev(new_gene_order), on='gene'][,
        list(subtype, mean_zscore), by='gene'])[order(-mean_zscore),
          gene[1:max_genes], by='subtype']$V1
  
  # best genes per cluster
  label_gene_list <- unique(unlist(label_genes, use.names=F))
  best_genes <- unique(c(best_genes, label_gene_list))
  
  #first 4 genes chosen per cluster
  glist <- list()
  for (i in 1:length(label_genes)) 
    glist[[i]] <- data.table(subtype=names(label_genes[i]), genes=label_genes[[i]][1:4])
  gene_top4 <- rbindlist(glist)
  
  # for annotation
  best_per_subtype <-unique(
    melted_genes[subtype == best_subtype][
      rev(new_gene_order), on='gene'][,
        list(subtype, mean(zscore)), by='gene'])[,
          is_best := gene %in% gene[1:max_genes], by='subtype']
  
  melted_genes[
    gene %in% best_genes &
    subtype == best_subtype &
    donor == donor[1],
      label := ifelse(gene %in% label_gene_list, as.character(gene), NA)]
    
  # alternate none with subtype name (two donors)
  row_names <- as.vector(t(cbind(rep('', length(subtype_order)), 
    cluster_name_dt[subtype_order, on='Subtype'][, subtype_lbl])))
  unique_donor_subset <- unique(melted_genes[, list(donor, subtype)])
  n_lines <- unique_donor_subset[, length(unique(subtype))]
  
  max_x <- length(levels(melted_genes[gene %in% best_genes, interaction(donor,subtype)]))
  
  display_genes <- melted_genes[gene %in% best_genes]
  
  separator_table <- display_genes[
      subtype == best_subtype & donor == donor[1], list(n_genes= as.double(.N)), by='subtype']
      
  separator_table[rev(subtype_order), on='subtype', pos := cumsum(n_genes)+0.5]
  separator_table[, x_pos := (n_lines-rank(pos))*2 + 0.5]
  
  label_table <- display_genes[,
    paste(na.omit(label), collapse=', '), by='subtype'][
      separator_table, on='subtype']
  
  top4_table <- gene_top4[,
    paste(na.omit(genes), collapse='\n'), by='subtype'][
      separator_table, on='subtype']
  
  de_gene_heatmap_boxed <- ggplot(display_genes) + 
    geom_tile(aes(x=interaction(donor,subtype), y=gene, fill=zscore)) +
    scale_x_discrete(expand=c(0,0)) + 
    scale_y_discrete(expand=c(0,0)) +
    geom_vline(color='white', xintercept=0.5+c(1:n_lines-1)*2) +
    geom_rect(data=separator_table,
      aes(ymax=pos-0.25, ymin=pos-n_genes+0.25, xmin=x_pos, color=subtype),
      xmax=max_x+label_width, fill=NA, size=0.6) +
    geom_rect(data=separator_table,
      aes(ymax=pos-0.25, ymin=pos-n_genes+0.25, xmin=x_pos, xmax=x_pos+2, color=subtype), fill=NA, size=0.6) +
    scale_color_manual(guide='none', values=unlist(all_colors)) +
    scale_fill_distiller('Gene\nz-score',
      palette='PuOr', oob=scales::squish, limit=c(-2, 2), type='div',
      breaks=c(2,0,-2), labels=c('≥2','0','≤-2')) +
    geom_label(
      data=top4_table,
      label.size=0,
      fill=NA,
      nudge_x=-0.55,
      size=label_font_size*1.8,
      aes(x=max_x+1, y=pos-n_genes, label=V1, vjust=-0.05, hjust=0)) +
    theme_minimal() +
    coord_flip(clip='off') +
    theme(
      axis.text.x=element_blank(),
      strip.text.y=element_blank(),
      axis.text.y=element_blank(),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(70, 5.5, 0, 5.5), "points"))
  
  donor_idents <- ggplot(unique_donor_subset) + 
    geom_tile(aes(y=interaction(donor, subtype), fill=donor, x='a')) + 
    scale_y_discrete(labels=row_names, expand=c(0,0)) + 
    scale_x_discrete(expand=c(0,0)) +
    geom_hline(color='white', yintercept=0.5+c(1:n_lines-1)*2) +
    scale_fill_discrete('Donor', labels=c('A','B')) + 
    theme_minimal() +
    coord_fixed(ratio=1) +
    theme(
      axis.text.y=element_blank(),
      axis.text.x=element_blank(),
      axis.title.y=element_blank(),
      axis.title.x=element_blank(),
      plot.margin = unit(c(22, 0, 5.5, 0), "points"))
  
  cluster_idents <- ggplot(unique_donor_subset) + 
    geom_tile(aes(y=interaction(donor, subtype), fill=subtype, x='a')) + 
    geom_hline(color='white', yintercept=0.5+c(1:n_lines-1)*2) +
    scale_y_discrete(labels=row_names, expand=c(0,0)) + 
    scale_x_discrete(expand=c(0,0)) +
    scale_fill_manual(values=all_colors) +
    theme_minimal() +
    coord_fixed(ratio=1) +
    theme(
      legend.position = 'none',
      axis.text.y = element_text(vjust = -0.5, hjust=1),
      axis.title.x= element_blank(),
      axis.title.y=element_blank(),
      axis.text.x=element_blank(),
      panel.background = element_blank(),
      plot.margin = unit(c(22, 0, 5.5, 5.5), "points"))
  

  return(cluster_idents | donor_idents | de_gene_heatmap_boxed)
}
```

### Assemble individual components of gene heatmaps

```{r}

if (!(exists('cd4_pruned'))) {
  Idents(scrna_cd4_all) <- 'Subtype'
  cd4_pruned <- prune_cells(scrna_cd4_all, rescale=T)
}

if (!(exists('cd8_pruned'))) {
  Idents(scrna_cd8_all) <- 'Subtype'
  cd8_pruned <- prune_cells(scrna_cd8_all, rescale=T)
}

# orig versions
plots$cd4_gene_heatmap <- make_gene_heatmap(
  cd4_pruned, cd4_order, 
  max_genes=50, label_genes=cd4_genes, rescale=F, 
  use_dendsort=T,
  label_width = 6.75, label_font_size=1.8, label_line_width=18) 

plots$cd8_gene_heatmap <- make_gene_heatmap(
  cd8_pruned, cd8_order, max_genes=50,
  label_genes=cd8_genes, rescale=F, 
  use_dendsort=T,
  label_width = 6.75, label_font_size=1.8, label_line_width=18)

plots$cd4_gene_heatmap[[1]] <- plots$cd4_gene_heatmap[[1]] + 
  scale_y_discrete('Top 50 DE genes per cluster') + ggtitle('CD4') + theme(
    legend.position='none', axis.title.y=element_text(angle = 90))
plots$cd4_gene_heatmap[[2]] <- plots$cd4_gene_heatmap[[2]] + theme(legend.position='none')

legend <- plot_grid(
  get_legend(
    # create some space to the left of the legend
    plots$cd8_gene_heatmap[[1]] + theme(
      legend.box.margin = margin(0, 0, 0, 12),
      legend.justification=c(1,1), legend.position=c(1,1),
      legend.key.width=unit(0.4, 'lines'))),
  get_legend(
    plots$cd8_gene_heatmap[[2]] + theme(
      legend.box.margin = margin(0, 0, 0, 12),
      legend.justification=c(1,0), legend.position=c(1,0),
      legend.key.width=unit(0.4, 'lines'))),
  ncol=1
)

plots$cd8_gene_heatmap[[1]] <- plots$cd8_gene_heatmap[[1]] + ggtitle('CD8')
plots$cd8_gene_heatmap[[1]] <- plots$cd8_gene_heatmap[[1]] + theme(legend.position='none')
plots$cd8_gene_heatmap[[2]] <- plots$cd8_gene_heatmap[[2]] + theme(legend.position='none')

heatmap_ratio = 2.2
plots$combined_gene_heatmap <- (
  plots$cd4_gene_heatmap | plot_spacer() | plots$cd8_gene_heatmap | plot_spacer() | (
  (plot_spacer() / legend) + plot_layout(heights=c(1,1)))) +
    plot_layout(widths=c(8,2,9,2,3))

plots$combined_gene_heatmap

# combined boxed

plots$cd4_gene_heatmap_boxed <- make_gene_heatmap(
  cd4_pruned, cd4_order, 
  max_genes=100, label_genes=cd4_genes, rescale=F, 
  use_dendsort=T, boxed=T,
  label_width = 7.25, label_font_size=1.5, label_line_width=18)

plots$cd8_gene_heatmap_boxed <- make_gene_heatmap(
  cd8_pruned, cd8_order, max_genes=50,
  label_genes=cd8_genes, rescale=F, 
  use_dendsort=T, boxed=T,
  label_width = 7.25, label_font_size=1.5, label_line_width=18)

plots$cd4_gene_heatmap_boxed[[1]] <- plots$cd4_gene_heatmap_boxed[[1]] + 
  scale_y_discrete('Top 50 DE genes per cluster') + ggtitle('CD4') + theme(
    legend.position='none', axis.title.y=element_text(angle = 90))
plots$cd4_gene_heatmap_boxed[[2]] <- plots$cd4_gene_heatmap_boxed[[2]] + theme(legend.position='none')

plots$cd8_gene_heatmap_boxed[[1]] <- plots$cd8_gene_heatmap_boxed[[1]] + ggtitle('CD8')
plots$cd8_gene_heatmap_boxed[[1]] <- plots$cd8_gene_heatmap_boxed[[1]] + theme(legend.position='none')
plots$cd8_gene_heatmap_boxed[[2]] <- plots$cd8_gene_heatmap_boxed[[2]] + theme(legend.position='none')

plots$combined_gene_heatmap_boxed <- (
  plots$cd4_gene_heatmap_boxed | plot_spacer() | plots$cd8_gene_heatmap_boxed | plot_spacer() | (
  (plot_spacer() / legend) + plot_layout(heights=c(1,1)))) +
    plot_layout(widths=c(8,2,9,2,3))

# combined boxed horiz

cd4_gene_heatmap_boxed_horiz_orig <- make_gene_heatmap_horiz(
  cd4_pruned, cd4_order, 
  max_genes=100,
  label_genes=cd4_genes, rescale=F, 
  use_dendsort=T, boxed=T,
  label_width = 8, label_font_size=1.5)

cd8_gene_heatmap_boxed_horiz_orig <- make_gene_heatmap_horiz(
  cd8_pruned, cd8_order,
  max_genes=100,
  label_genes=cd8_genes, rescale=F, 
  use_dendsort=T, boxed=T,
  label_width = 8, label_font_size=1.5)
  
  
plots$cd4_gene_heatmap_boxed_horiz <- cd4_gene_heatmap_boxed_horiz_orig + 
  theme(plot.margin = unit(c(35, 0, 0, 5.5), "points"))

plots$cd8_gene_heatmap_boxed_horiz <- cd8_gene_heatmap_boxed_horiz_orig + 
  theme(plot.margin = unit(c(35, 0, 0, 5.5), "points"))

plots$cd4_gene_heatmap_boxed_horiz[[1]] <- plots$cd4_gene_heatmap_boxed_horiz[[1]] + 
  ggtitle('CD4')
plots$cd4_gene_heatmap_boxed_horiz[[3]] <- plots$cd4_gene_heatmap_boxed_horiz[[3]] +
  theme(legend.position='none')
plots$cd4_gene_heatmap_boxed_horiz[[2]] <- plots$cd4_gene_heatmap_boxed_horiz[[2]] + 
  theme(legend.position='none')

legend_horiz <- plot_grid(
  get_legend(
    # create some space to the left of the legend
    plots$cd8_gene_heatmap_boxed_horiz[[3]] + theme(
      legend.box.margin = margin(0, 0, 0, 12),
      legend.justification=c(1,1), legend.position=c(1,1),
      legend.key.width=unit(0.4, 'lines'))),
  get_legend(
    plots$cd8_gene_heatmap_boxed_horiz[[2]] + theme(
      legend.box.margin = margin(0, 0, 0, 12),
      legend.justification=c(1,0), legend.position=c(1,0),
      legend.key.width=unit(0.4, 'lines'))),
  ncol=1
)

plots$cd8_gene_heatmap_boxed_horiz[[1]] <- plots$cd8_gene_heatmap_boxed_horiz[[1]] + 
  ggtitle('CD8')
plots$cd8_gene_heatmap_boxed_horiz[[3]] <- plots$cd8_gene_heatmap_boxed_horiz[[3]] + 
  labs(y='Top 100 DE genes per cluster') + theme(axis.title.x=element_text()) 
plots$cd8_gene_heatmap_boxed_horiz[[2]] <- plots$cd8_gene_heatmap_boxed_horiz[[2]] + theme(legend.position='none')

plots$combined_gene_heatmap_boxed_horiz <- ((
  (plots$cd4_gene_heatmap_boxed_horiz / plots$cd8_gene_heatmap_boxed_horiz) + 
    plot_layout(heights=c(8,9)) & theme(legend.position='none')) |
  legend_horiz) + plot_layout(widths=c(10,1)) & 
  theme(
    plot.title.position = "plot",
    plot.title=element_text(margin=unit(c(0,0,0,0),'points'), hjust=0))

```


### ADT


```{r}
adt_groups <- list(
  'Activation/Costimulation'= c(
    "IL2RA",
    "CD137",
    "OX40",
    "CD30",
    "CD80",
    "CD86",
    "CD38",
    "ICOS",
    "GITR",
    "CD2",
    "CD69",
    "CD40",
    "HLA-ABC"),
  "Cytokines/Chemokines\nIntegrins"=c(
    "IL2RA",
    "IL7R",
    "CXCR3",
    "CXCR5",
    "CD49E",
    "ITGA1",
    "CD49b",
    "CD49E",
    "CCR5"),
  "Differentiation"=c(
    "CD45RO",
    "CD45RA",
    "CCR7",
    "CD62L",
    "CD7",
    "CD95",
    "CD27",
    "CD28"),
  "Inhibitory"=c(
    "PD1",
    "LAG3",
    "Tim3"))


make_adt_heatmap <- function(seurat_obj, adt_groups, subtype_order) {
  
  seurat_obj <- ScaleData(seurat_obj, assay='SCT_ADT_INT')
  
  all_adt <- AverageExpression(
    seurat_obj,
    assays = 'SCT_ADT_INT',
    features = NULL,
    return.seurat = FALSE,
    group.by = "Subtype",
    slot = "scale.data"
  )

  # melt, clean
  melted_adt <- reshape2::melt(all_adt[[1]])
  melted_adt <- data.table(melted_adt)
  names(melted_adt) <- c('gene','subtype','zscore')
  melted_adt[, donor := gsub('.*_(.*)','\\1', subtype)]
  melted_adt[, subtype := gsub('(.*)_(.*)','\\1', subtype)]
  melted_adt[, subtype := factor(subtype, levels=subtype_order)]
  melted_adt[, best_subtype := subtype[which.max(zscore)], by=gene]
  melted_adt[, zdiff := zscore[which.max(zscore)] - mean(zscore), by=gene]
  gene_order <- unique(melted_adt[subtype == best_subtype, gene[order(zdiff)]])
  melted_adt[, gene := factor(gene, levels=gene_order)]
  melted_adt[, max_score := max(abs(zscore)), by='gene']
  
  # remove '.adt'
  melted_adt[, gene := gsub('\\.adt','',gene)]
  

  # assign groups
  if (!is.null(adt_groups)) {
    for(i in 1:length(adt_groups)) {
      melted_adt[gene %in% adt_groups[[i]], group := names(adt_groups)[i]]
    }
    #reorder
    new_adt_order <- c()
    for (group_i in na.omit(unique(melted_adt$group))) {
      
      cast_group <- dcast(melted_adt[!is.na(subtype) & group == group_i], 
          gene ~ subtype, value.var='zscore')
      
      group_adt_order <- dendsort(hclust(dist(
        dcast(melted_adt[group == group_i], 
          gene ~ subtype, value.var='zscore')[,-1])))$order
      group_adts <- cast_group$gene[group_adt_order]
      new_adt_order <- c(new_adt_order, group_adts)
    }
    melted_adt <- melted_adt[!is.na(group)]
    melted_adt[, gene := factor(gene, levels=new_adt_order)]
    cast_formula <- "group~."
  } else {
    cast_formula <- "best_subtype ~."
  }
  
  de_adt_heatmap <- ggplot(melted_adt) + 
    geom_raster(aes(x=factor(subtype,
      levels=cluster_name_dt[subtype_order, on='Subtype']$Subtype,
      labels=cluster_name_dt[subtype_order, on='Subtype']$subtype_lbl),
      y=gene, fill=zscore)) +
    facet_grid(as.formula(cast_formula), scales='free', space='free') +
    scale_fill_distiller(
      'CITE-seq z-score', palette='PuOr',
      oob=scales::squish, limit=c(-1, 1), type='div', breaks=c(-1,0,1),
      labels=c('>1', '0', '<1')) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))
  
  return(de_adt_heatmap)
}

make_adt_heatmap <- function(seurat_obj, adt_groups, subtype_order, return_data=F) {
  
  seurat_obj <- ScaleData(seurat_obj, assay='SCT_ADT_INT')
  
  all_adt <- AverageExpression(
    seurat_obj,
    assays = 'SCT_ADT_INT',
    features = NULL,
    return.seurat = FALSE,
    group.by = "Subtype",
    slot = "scale.data"
  )

  # melt, clean
  melted_adt <- reshape2::melt(all_adt[[1]])
  melted_adt <- data.table(melted_adt)
  names(melted_adt) <- c('gene','subtype','zscore')
  melted_adt[, donor := gsub('.*_(.*)','\\1', subtype)]
  melted_adt[, subtype := gsub('(.*)_(.*)','\\1', subtype)]
  melted_adt[, subtype := factor(subtype, levels=subtype_order)]
  melted_adt[, best_subtype := subtype[which.max(zscore)], by=gene]
  melted_adt[, zdiff := zscore[which.max(zscore)] - mean(zscore), by=gene]
  gene_order <- unique(melted_adt[subtype == best_subtype, gene[order(zdiff)]])
  melted_adt[, gene := factor(gene, levels=gene_order)]
  melted_adt[, max_score := max(abs(zscore)), by='gene']
  
  # remove '.adt'
  melted_adt[, gene := gsub('\\.adt','',gene)]
  

  # assign groups
  if (!is.null(adt_groups)) {
    for(i in 1:length(adt_groups)) {
      melted_adt[gene %in% adt_groups[[i]], group := names(adt_groups)[i]]
    }
    #reorder
    new_adt_order <- c()
    for (group_i in na.omit(unique(melted_adt$group))) {
      
      cast_group <- dcast(melted_adt[!is.na(subtype) & group == group_i], 
          gene ~ subtype, value.var='zscore')
      
      group_adt_order <- dendsort(hclust(dist(
        dcast(melted_adt[group == group_i], 
          gene ~ subtype, value.var='zscore')[,-1])))$order
      group_adts <- cast_group$gene[group_adt_order]
      new_adt_order <- c(new_adt_order, group_adts)
    }
    melted_adt <- melted_adt[!is.na(group)]
    melted_adt[, gene := factor(gene, levels=new_adt_order)]
    cast_formula <- "group~."
  } else {
    cast_formula <- "best_subtype ~."
  }
  
  de_adt_heatmap <- ggplot(melted_adt) + 
    geom_raster(aes(x=factor(subtype,
      levels=cluster_name_dt[subtype_order, on='Subtype']$Subtype,
      labels=cluster_name_dt[subtype_order, on='Subtype']$subtype_lbl),
      y=gene, fill=zscore)) +
    facet_grid(as.formula(cast_formula), scales='free', space='free') +
    scale_fill_distiller(
      'CITE-seq z-score', palette='PuOr',
      oob=scales::squish, limit=c(-1, 1), type='div', breaks=c(-1,0,1),
      labels=c('>1', '0', '<1')) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
      axis.title.x=element_blank(),
      axis.title.y=element_blank(),
      plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))
  
  if (return_data) {
    melted_adt
  } else {
    return(de_adt_heatmap)
  }
}

cd8_stim_subsets <- c('STIM-GLYC','STIM-OXPHOS','STIM-IFNG','STIM-GNLY','STIM-INACT','STIM-UNTR')
cd4_stim_subsets <- c('STIM-GLYC','STIM-OXPHOS','STIM-IFNG','STAT1-IRF1','STIM-UNTR')

cd8_stim_prune <- prune_cells(scrna_cd8_all, subtypes= cd8_stim_subsets, k_types='cd19+')
cd8_stim_citeseq_map <- make_adt_heatmap(cd8_stim_prune, adt_groups, cd8_order)

cd4_stim_prune <- prune_cells(scrna_cd4_all, subtypes= cd4_stim_subsets, k_types='cd19+')
cd4_stim_citeseq_map <- make_adt_heatmap(cd4_stim_prune, adt_groups, cd4_order)

combined_adt_data <- rbind(make_adt_heatmap(cd8_stim_prune, adt_groups, cd8_order, return_data=T)[, t_type := 'CD8'], make_adt_heatmap(cd4_stim_prune, adt_groups, cd4_order,  return_data=T)[, t_type := 'CD4'])

plots$combined_adt_plot <- ggplot(combined_adt_data) + 
  geom_raster(aes(x=factor(subtype,
    levels=cluster_name_dt[comb_order, on='Subtype']$Subtype,
    labels=cluster_name_dt[comb_order, on='Subtype']$subtype_lbl),
    y=gene, fill=zscore)) +
  facet_grid('group~t_type', scales='free', space='free') +
  scale_x_discrete(expand=c(0,0)) +
  scale_y_discrete(expand=c(0,0)) +
  scale_fill_distiller(
    'CITE-seq z-score', palette='PuOr',
    oob=scales::squish, limit=c(-1, 1), type='div', breaks=c(-1,0,1),
    labels=c('>1', '0', '<1')) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
    axis.title.x=element_blank(),
    axis.title.y=element_blank(),
    panel.border=element_rect(color='grey70', fill=NA),
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))

rm(cd8_stim_prune)
rm(cd4_stim_prune)

```


```{r}
# "NAIVE-CD7"=
#  "IFITM2, VIM, STMN1, HMGN2, CAPG, STIM1, CD7, IFITM1, PLEKHO1, TCF7"
# "MEM-CD29"=
#  "SELL, ZEB2, EOMES, ITGA1, ITGB2, ITGB7, ITGA4, ITGB1, PLEK, CXCR3, CD59, ID2, IGFBP3, MHC-II, PYHIN1, PRR5L, S100A4, ANXA1, CCL5, TNFSF9, BATF, CX3CR1, FGFBP2, KLRG1, NKG7, GZMK, CST7, HOPX"
# "STIM-IFNG"=
#  "IFNG, GZMH, GZMK, GZMB, NKG7, CCR5, HLA-DQA1, CD74, PRDM1, CD2, LAG3, TNFRSF9, FABP5, CCL5, FAM162A, LGALS1, SRM, HLA-DRA, HLA-DQA2, ID2, GNLY, GZMB, ZBED2, ALDOC, HOPX, CTSD, ENO1, SLAMF1, DPP4, FCGR3A, PYHIN1"
# "STIM-GLYC"=
#  "ARG2, BNIP3, BNIP3L, LAYN, HILPDA, MT2, MT2A, MT1E, MT1X, LDHA, DDIT3, DDIT4, FAM162A, PKM, JUNB, XCL2, MTHFD2, DUSP4, LDHA, ALDOC, SLC2A3, LTA, GADD45A, JUND, TNFAIP3"
# "STIM-UNTR"=
#  "S100A10, CD38, LAYN, ENO1, PGK1, IGFBP2, NEAT1, STAT1, CD27, TNFAIP2"
# "STIM-OXPHOS"=
#  "HMGA1, HSP90AA1, FKBP4 (mtor), HSPD1, MT-CO3, HSP90AB1, CXCL10, IL5, C1QBP, EIF5A, HSPH1, MYC, MD-ND3, SRM, MTHFD2, MT-ATP6, HSPE1, HSP8, IL22, IL13, IL2RA"
# "STIM-INACT"=
#  "BATF, CTSW, DDX21, CYP1B1, BATF3, GZMA, TNFSF10, PFN1"
# "NAIVE-CD62L"=
#  "SELL, MAL, KLF2, TXNIP, IL7R, LRRN3"
# "MEM-CD29"=
#  "CD8A, CXCR3, CD28, CD45RO, CD29, HVEM"
# "STIM-IFNG"=
#  "CD2, CD49b, CD45RO, CD28"
# "STIM-GLYC"=
#  "IL2RA, PD1, LAG3"
# "STIM-UNTR"=
#  "CD5, CCR5, CCR7"
# "STIM-OXPHOS"=
#  "ICOS, CD2, LAG3, CD80, CD30, ICOS, IL2RA"
# "STIM-INACT"=
#  "Tim3, CD2, CD7, CCR6, TNFRSF18"

# NAIVE-CD7: IFITM2, VIM, STMN1, HMGN2, CAPG, STIM1, CD7, IFITM1, PLEKHO1, TCF7
# 
# MEM-CD29_genes: SELL, ZEB2, EOMES, ITGA1, ITGB2, ITGB7, ITGA4, ITGB1, PLEK, CXCR3, CD59, ID2, IGFBP3, MHC-II, PYHIN1, PRR5L, S100A4, ANXA1, CCL5, TNFSF9, BATF, CX3CR1, FGFBP2, KLRG1, NKG7, GZMK, CST7, HOPX
# MEM-CD29-proteins: CD8A, CXCR3, CD28, CD45RO, CD29, HVEM
# 
# STIM-IFNG-genes: PRDM1, CCR5, HLA-DQA1, CD2, LAG3, TNFRSF9, PRDM1, FABP5, CCL5, FAM162A, LGALS1, SRM, IFNG, HLA-DRA, HLA-DQA2, GZMK, ID2, GNLY, GZMB, ZBED2, ALDOC, HOPX, CTSD, ENO1, SLAMF1, DPP4, FCGR3A, PYHIN1, NKG7, CD74, GZMH, LTB, PLEK, ANXA1, KLRG1
# STIM-IFNG-proteins: CD2, CD49b, CD45RO, CD28
# 
# STIM-GLYC-genes: ARG2, BNIP3, BNIP3L, LAYN, HILPDA, MT2, MT2A, MT1E, MT1X, LDHA, DDIT3, DDIT4, FAM162A, PKM, JUNB, XCL2, MTHFD2, DUSP4, LDHA, ALDOC, SLC2A3, LTA, GADD45A, JUND, TNFAIP3
# STIM-GLYC-proteins: IL2RA, PD1, LAG3
# 
# STIM-UNTR-genes: S100A10, CD38, LAYN, ENO1, PGK1, IGFBP2, NEAT1, STAT1, CD27, TNFAIP2
# STIM-UNTR-proteins: CD5, CCR5, CCR7
# 
# STIM-OXPHOS-genes: HMGA1, HSP90AA1, FKBP4 (mtor), HSPD1, MT-CO3, HSP90AB1, CXCL10, IL5,
# C1QBP, EIF5A, HSPH1, MYC, MD-ND3, SRM, MTHFD2, MT-ATP6, HSPE1, HSP8, IL22, IL13, IL2RA
# STIM-OXPHOS-proteins: ICOS, CD2, LAG3, CD80, CD30
# 
# STIM-INACT-genes: BATF, CTSW, DDX21, CYP1B1, BATF3, GZMA, TNFSF10, PFN1
# STIM-INACT-proteins: Tim3, CD2, CD7, CCR6, GITR

```
### ADT Umap plots

```{r}
umap_proteins <- c(
  'CD62L',
  'CD45RA',
  'CD45RO',
  'HLA-ABC',
  'CD49b',
  'IL2RA')

umap_protein_ranges <- list(
  'CD62L.adt'=c(0,4),
  'CD45RA.adt'=c(0,3),
  'CD45RO.adt'=c(-0.5,1.75),
  'HLA-ABC.adt'=c(0,2.5),
  'CD49b.adt'=c(0,2.5),
  'IL2RA.adt'=c(0,8))

umap_range_dt <- data.table(
  variable=names(umap_protein_ranges),
  var.min=sapply(umap_protein_ranges, function (x) x[1]),
  var.max=sapply(umap_protein_ranges, function (x) x[2]))

prev_assay <- DefaultAssay(scrna_all)
DefaultAssay(scrna_all) <- 'SCT_ADT_INT'
citeseq_data <- data.table(
  FetchData(
    scrna_all,
    names(umap_protein_ranges),
    slot='data'),
  keep.rownames=T)
DefaultAssay(scrna_all) <- prev_assay
setnames(citeseq_data, 'rn','X_index')
citeseq_umap <- citeseq_data[umap_dt, on='X_index']
csu_melt <- melt(citeseq_umap, measure.vars=paste0(umap_proteins,'.adt'))

csu_melt <- csu_melt[umap_range_dt, on='variable'][, 
  value.scaled := scales::rescale(
    scales::oob_squish(value, range=c(var.min[1], var.max[1]))),
  by='variable']

var_order <- c('CD62L','CD45RA','CD45RO','CD49b','HLA-ABC','IL2RA')
csu_melt <- csu_melt[,
  variable := factor(variable, levels=paste0(var_order,'.adt'), labels=var_order)]

plots$citeseq_umaps <- ggplot(csu_melt,
    aes(x=wnnUMAP_1, y=wnnUMAP_2, color=value.scaled)) +
  rasterise(geom_point(size=inset_umap_point_size, alpha=0.5), dpi=500) + 
  scale_color_distiller('Protein expression (CITE-Seq)',
    palette='RdPu', direction=1, limit=c(0,1),
    breaks=c(0,1),
    labels=c('min','max')) +
  coord_fixed() +
  facet_wrap(~variable) +
  theme_minimal() + 
  theme(
    legend.position='bottom',
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())
```

### RNA Umap plots (supp)

```{r}
umap_rna_ranges <- list(
  'IFNG'=c(0,6),
  'KLRB1'=c(0,10),
  'GZMB'=c(0,15),
  'HLA-DRA'=c(0,7),
  'ITGB1'=c(0,5),
  'EOMES'=c(0,4),
  'S100A4'=c(0,5),
  'PRF1'=c(0,3))

umap_rna_range_dt <- data.table(
  variable=names(umap_rna_ranges),
  var.min=sapply(umap_rna_ranges, function (x) x[1]),
  var.max=sapply(umap_rna_ranges, function (x) x[2]))

prev_assay <- DefaultAssay(scrna_all)
DefaultAssay(scrna_all) <- 'SCT_INT'
umap_rna_data <- data.table(
  FetchData(
    scrna_all,
    names(umap_rna_ranges),
    slot='data'),
  keep.rownames=T)
DefaultAssay(scrna_all) <- prev_assay
setnames(umap_rna_data, 'rn','X_index')
sct_rna_umap <- umap_rna_data[umap_dt, on='X_index']
csu_melt <- melt(sct_rna_umap, measure.vars=names(umap_rna_ranges))

csu_melt <- csu_melt[umap_rna_range_dt, on='variable'][, 
  value.scaled := scales::rescale(
    scales::oob_squish(value, range=c(var.min[1], var.max[1]))),
  by='variable']

var_order <-  names(umap_rna_ranges)
csu_melt <- csu_melt[,
  variable := factor(variable, levels=var_order)]

plots$rna_umaps <- ggplot(csu_melt,
    aes(x=wnnUMAP_1, y=wnnUMAP_2, color=value.scaled)) +
  rasterise(geom_point(size=inset_umap_point_size, alpha=0.5), dpi=100) + 
  scale_color_distiller('RNA expression (SCTransform)',
    palette='RdPu', direction=1, limit=c(0,1),
    breaks=c(0,1),
    labels=c('min','max')) +
  coord_fixed() +
  facet_wrap(~variable, nrow=1) +
  theme_minimal() + 
  theme(
    legend.position='bottom',
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())

plots$rna_umaps_rect <- ggplot(csu_melt,
    aes(x=wnnUMAP_1, y=wnnUMAP_2, color=value.scaled)) +
  rasterise(geom_point(size=inset_umap_point_size, alpha=0.5), dpi=100) + 
  scale_color_distiller('RNA expression (SCTransform)',
    palette='RdPu', direction=1, limit=c(0,1),
    breaks=c(0,1),
    labels=c('min','max')) +
  coord_fixed() +
  facet_wrap(~variable, nrow=2, ncol=4) +
  theme_minimal() + 
  theme(
    legend.position='bottom',
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    legend.justification = c(0,1),
    panel.grid=element_blank())
```

## Pathways

```{r}
if (!(exists('cd4_pruned'))) {
  Idents(scrna_cd4_all) <- 'Subtype'
  cd4_pruned <- prune_cells(scrna_cd4_all, rescale=T)
}

if (!(exists('cd8_pruned'))) {
  Idents(scrna_cd8_all) <- 'Subtype'
  cd8_pruned <- prune_cells(scrna_cd8_all, rescale=T)
}

gsea_subsets <- c(
  "HALLMARK_PI3K_AKT_MTOR_SIGNALING",
  "KEGG_MTOR_SIGNALING_PATHWAY",
  "HALLMARK_MTORC1_SIGNALING",
  "BIOCARTA_NFKB_PATHWAY",
  "GILMORE_CORE_NFKB_PATHWAY",
  "HALLMARK_GLYCOLYSIS",
  "HALLMARK_OXIDATIVE_PHOSPHORYLATION",
  "HALLMARK_IL6_JAK_STAT3_SIGNALING",
  "HALLMARK_MYC_TARGETS_V2",
  "BIOCARTA_CD40_PATHWAY",
  "BIOCARTA_41BB_PATHWAY",
  "FATTY_ACID_OXIDATION",
  "HALLMARK_HYPOXIA",
  "GOBP_INTERFERON_GAMMA_MEDIATED_SIGNALING_PATHWAY",
  "GOBP_ACTIVATION_OF_NF_KAPPAB_INDUCING_KINASE_ACTIVITY",
  "GOBP_NEGATIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",
  "GOBP_NEGATIVE_REGULATION_OF_TYPE_I_INTERFERON_MEDIATED_SIGNALING_PATHWAY",
  "BIOCARTA_IFNG_PATHWAY",
  "PID_IL12_2PATHWAY",
  "HALLMARK_APOPTOSIS",
  "HALLMARK_INTERFERON_GAMMA_RESPONSE",
  "KEGG_CELL_ADHESION_MOLECULES_CAMS",
  "REACTOME_INTERFERON_GAMMA_SIGNALING",
  "HALLMARK_IL2_STAT5_SIGNALING",
  "PID_CD40_PATHWAY",
  "PID_NFAT_TFPATHWAY",
  "GOBP_ACTIVATION_INDUCED_CELL_DEATH_OF_T_CELLS",
  "PID_IL2_PI3K_PATHWAY",
  "REACTOME_FASL_CD95L_SIGNALING",
  "PID_FAS_PATHWAY",
  "WP_TNF_ALPHA_SIGNALING_PATHWAY",
  "BIOCARTA_IFNA_PATHWAY",
  "WP_TYPE_II_INTERFERON_SIGNALING_IFNG",
  "BIOCARTA_NKT_PATHWAY",
  "WP_ARYL_HYDROCARBON_RECEPTOR_PATHWAY",
  "REACTOME_INTERLEUKIN_12_SIGNALING",
  "PID_SHP2_PATHWAY",
  "BIOCARTA_CTL_PATHWAY",
  "BOSCO_TH1_CYTOTOXIC_MODULE",
  "PID_NFAT_TFPATHWAY",
  "GOBP_TYPE_I_INTERFERON_PRODUCTION",
  "KEGG_CALCIUM_SIGNALING_PATHWAY",
  "BIOCARTA_GATA3_PATHWAY",
  "PID_AP1_PATHWAY",
  "MOOTHA_FFA_OXYDATION",
  "MOOTHA_GLUCONEOGENESIS",
  "MOOTHA_GLYCOGEN_METABOLISM",
  "MOOTHA_GLYCOLYSIS",
  "MOOTHA_ROS",
  "MOOTHA_TCA",
  "MOOTHA_VOXPHOS"
  )

# look at DE genes in each pathway:
#combined_de_markers[rn %in% get_geneset('MOOTHA_FFA_OXYDATION')][order(avg_log2FC)]

gsea.pathways.file <- here::here(
     '..','..','scrnaseq_tcsl154',
     'rds',
     paste0('gsea_pathways.Rdata'))

if (rerun_geneset_scores) {
  cd8_gsea_scores_rna <- map_geneset_parallel(cd8_pruned, gsea_subsets, assay='RNA', nbin=24)
  cd8_pruned[[names(cd8_gsea_scores_rna)]] <- cd8_gsea_scores_rna
  
  cd8_all_gsea_scores_rna <- map_geneset_parallel(scrna_cd8_all, gsea_subsets, assay='RNA', nbin=24)
  scrna_cd8_all[[names(cd8_all_gsea_scores_rna)]] <- cd8_all_gsea_scores_rna
  
  cd4_all_gsea_scores_rna <- map_geneset_parallel(scrna_cd4_all, gsea_subsets, assay='RNA', nbin=24)
  scrna_cd4_all[[names(cd4_all_gsea_scores_rna)]] <- cd4_all_gsea_scores_rna
  
  cd4_gsea_scores_rna <- map_geneset_parallel(cd4_pruned, gsea_subsets, assay='RNA', nbin=16)
  cd4_pruned[[names(cd4_gsea_scores_rna)]] <- cd4_gsea_scores_rna
  
  save(list=c(
    'cd8_gsea_scores_rna',
    'cd8_all_gsea_scores_rna',
    'cd4_gsea_scores_rna',
    'cd4_all_gsea_scores_rna'),
    file=gsea.pathways.file)
} else {
  if (!exists('cd8_gsea_scores_rna'))
    load(gsea.pathways.file)
  cd8_pruned[[names(cd8_gsea_scores_rna)]] <- cd8_gsea_scores_rna
  cd4_pruned[[names(cd4_gsea_scores_rna)]] <- cd4_gsea_scores_rna
  scrna_cd8_all[[names(cd8_all_gsea_scores_rna)]] <- cd8_all_gsea_scores_rna
  scrna_cd4_all[[names(cd4_all_gsea_scores_rna)]] <- cd4_all_gsea_scores_rna
}

make_cdf <- function(seurat_obj, pathway_name, pathway_label, 
    subtypes=c('STIM-IFNG','STIM-GLYC','STIM-OXPHOS'), quant_limit=0.05, group.by='Subtype', colorlist=unlist(all_colors), return_val='cdf') {
  
  # pull out pathway
  data <- data.table(FetchData(seurat_obj, 
      unique(c(pathway_name,'Subtype','car', group.by))))

  # subset on subtypes
  if (!is.null(subtypes)) {
    data <- data[Subtype %in% subtypes]
  }
  
  data[, c(group.by) := factor(get(group.by), levels=unique(get(group.by)))]
  
  # rescale z-scores within groups
  data[, c(pathway_name) := list(scale(get(pathway_name)))]
  
  # get quantiles for plotting
  min_quant <- data[, quantile(get(pathway_name), c(quant_limit), na.rm=T), by=group.by][, min(V1)]
  max_quant <- data[, quantile(get(pathway_name), c(1-quant_limit), na.rm=T), by=group.by][, max(V1)]

  test_results <- data.table()
  
  #ks_test
  for (group in levels(data[, get(group.by)])) {
    koltest <- ks.test(
        data[get(group.by) == group, get(pathway_name)],
        data[get(group.by) != group, get(pathway_name)], exact=T)
    test_results <- rbind(
      test_results,
      data.table(
        'p.value'=koltest$p.value, 
        'statistic'=koltest$statistic, 
        'mean'=data[get(group.by) == group, mean(get(pathway_name))],
        'group'=group))
  }
  
  return(switch(return_val,
   'cdf'= ggplot(data,
      aes(x=get(pathway_name), color=get(group.by))) + 
      stat_ecdf(geom='step') +
      theme_minimal() +
      scale_y_continuous(breaks=c(0,1)) +
      scale_color_manual(values=colorlist) +
      theme(legend.position = 'none', panel.grid = element_blank()) +
      geom_vline(xintercept=-Inf) +
      coord_cartesian(xlim=c(min_quant, max_quant)) +
      geom_hline(yintercept=-Inf) +
      labs(x=pathway_label, y=''),
   'pdf'= ggplot(data,
      aes(x=get(pathway_name), color=get(group.by))) + 
      geom_density() +
      theme_minimal() +
      scale_y_continuous(breaks=c(0,1)) +
      scale_color_manual(values=colorlist) +
      theme(legend.position = 'none', panel.grid = element_blank()) +
      geom_vline(xintercept=-Inf) +
      coord_cartesian(xlim=c(min_quant, max_quant)) +
      geom_hline(yintercept=-Inf) +
      labs(x=pathway_label, y=''),
    'test'= test_results))
}



pathway_list <- list(
  'IFNG'='REACTOME_INTERFERON_GAMMA_SIGNALING',
  'NFkB'='BIOCARTA_NFKB_PATHWAY',
  'Th1/Cytotoxic'='BOSCO_TH1_CYTOTOXIC_MODULE',
  'Ca2+ Signaling'='KEGG_CALCIUM_SIGNALING_PATHWAY',
  'Cell Adhesion'='KEGG_CELL_ADHESION_MOLECULES_CAMS',
  'OXPHOS'='HALLMARK_OXIDATIVE_PHOSPHORYLATION',
  'TCA'="MOOTHA_TCA",
  'Th2/GATA3'="BIOCARTA_GATA3_PATHWAY",
  'STAT5'="HALLMARK_IL2_STAT5_SIGNALING",
  'Glycolysis'='HALLMARK_GLYCOLYSIS',
  'HIF1a/Hypoxia'='HALLMARK_HYPOXIA',
  'AICD'='GOBP_ACTIVATION_INDUCED_CELL_DEATH_OF_T_CELLS',
  'Myc'='HALLMARK_MYC_TARGETS_V2',
  'IL2/PI3K'='PID_IL2_PI3K_PATHWAY'
)

# plot_grid(
#   plotlist=lapply(1:length(pathway_list), function(i) {
#     make_cdf(cd4_pruned, pathway_list[[i]], pathway_label=names(pathway_list)[i])
#   }),
#   nrow=1,
#   align='hv',
#   axis='lr')
# 
# plot_grid(
#   plotlist=lapply(1:length(pathway_list), function(i) {
#     make_cdf(cd8_pruned, pathway_list[[i]], pathway_label=names(pathway_list)[i])
#   }),
#   nrow=1,
#   align='hv',
#   axis='lr')
# 
# plot_grid(
#   plotlist=lapply(1:length(pathway_list), function(i) {
#     make_cdf(cd8_pruned, pathway_list[[i]], pathway_label=names(pathway_list)[i], group.by='car', colorlist=car_colors)
#   }),
#   nrow=1,
#   align='hv',
#   axis='lr')

module_tile_grid <- function(seurat_obj, pathway_list, group.by, dots=F,
    subtypes=c('STIM-IFNG','STIM-GLYC','STIM-OXPHOS'),
    override_pathway_order=NULL, override_group_order=NULL) {
  
  test_table <- rbindlist(lapply(1:length(pathway_list), function(i) {
      test <- make_cdf(seurat_obj, pathway_list[[i]], 
        pathway_label=names(pathway_list)[i], subtypes=subtypes,
        group.by=group.by, return_val='test')
      test$pathway <- pathway_list[[i]]
      test$pathway_name <- names(pathway_list)[i]
      return(test)
      }))
  
  # remove duplicated pathways (I dont want to re-run addmodule)
  test_table <- unique(test_table)
  
  test_table[, pathway_name := factor(pathway_name)]
  

  pathway_order <- dendsort(hclust(t(dist(
    dcast(test_table, 
      pathway_name ~ group, value.var='mean')[,-1]))))$order
  group_order <- dendsort(hclust(t(dist(
    dcast(test_table, 
      group ~ pathway_name, value.var='mean')[,-1]))))$order
  if (!is.null(override_pathway_order))
    pathway_order <- override_pathway_order
  if (!is.null(override_group_order))
    group_order <- override_group_order
  
  test_table[, pathway_name := factor(pathway_name, levels=levels(pathway_name)[pathway_order])]
  test_table[, group := factor(group, levels=levels(factor(group))[group_order])]
  
  if (dots) {
    
    plot <- ggplot(test_table) + 
      geom_point(aes(
        y=group,
        x=pathway_name, color=mean, size=pmax(pmin(-log(p.value), 12), 0))) +
      scale_color_distiller('Module z-score',
        palette='PuOr', oob=scales::squish, limit=c(-1, 1), type='div',
        breaks=c(1, 0, -1), labels=c('>1','0','< -1')) + 
      scale_x_discrete('Gene Modules') +
      scale_y_discrete('') +
      theme_minimal() +
      theme(
        panel.border=element_rect(color='black', fill=NA),
        panel.grid = element_blank(),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
      coord_flip()
  } else {
  
    plot <- ggplot(test_table) +
      geom_tile(aes(
        y=group,
        x=pathway_name, fill=mean)) +
      scale_fill_distiller('Module z-score',
        palette='PuOr', oob=scales::squish, limit=c(-1, 1), type='div',
        breaks=c(1, 0, -1), labels=c('>1','0','< -1')) + 
      scale_x_discrete('Gene Modules', expand=c(0,0)) +
      scale_y_discrete('', expand=c(0,0)) +
      theme_minimal() +
      theme(
        panel.border=element_rect(color='black', fill=NA),
        legend.key.width=unit(0.4, 'lines'),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
      coord_flip()
  }
  
  return(list(
    data=test_table, plot=plot, pathway_order=pathway_order, group_order=group_order))
}

# ACTIVATED SUBSET PATHWAYS
activated_subtypes <- c('STIM-GLYC','STIM-OXPHOS','STIM-IFNG')

cd8_pruned[['subtype_lbl']] <- cluster_name_dt[cd8_pruned[['Subtype']], on='Subtype']$subtype_lbl
plots$cd8_module_tiles <- module_tile_grid(cd8_pruned,
    pathway_list, group.by='subtype_lbl', dots=F, subtypes=activated_subtypes)

cd4_pruned[['subtype_lbl']] <- cluster_name_dt[cd4_pruned[['Subtype']], on='Subtype']$subtype_lbl
plots$cd4_module_tiles <- module_tile_grid(cd4_pruned,
    pathway_list, group.by='subtype_lbl', dots=F, 
  subtypes=activated_subtypes,
  override_pathway_order = plots$cd8_module_tiles$pathway_order,
  override_group_order = plots$cd8_module_tiles$group_order)

#combined cd4/cd8 plot:

combined_tile_data <- rbind(
  data.table(plots$cd8_module_tiles$data)[, t_type := 'CD8'],
  data.table(plots$cd4_module_tiles$data[, t_type := 'CD4']))

plots$combined_tile_plot <- ggplot(combined_tile_data) +
      geom_tile(aes(
        y=group,
        x=pathway_name, fill=mean)) +
      scale_fill_distiller('Module\nz-score',
        palette='PuOr', oob=scales::squish, limit=c(-1, 1), type='div',
        breaks=c(1, 0, -1), labels=c('>1','0','< -1')) + 
      scale_x_discrete('Gene Modules', expand=c(0,0)) +
      scale_y_discrete('', expand=c(0,0)) +
      facet_grid(~t_type) +
      theme_minimal() +
      theme(
        panel.border=element_rect(color='black', fill=NA),
        legend.key.width=unit(0.4, 'lines'),
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
      coord_flip()

plots$cd8_expanded_modules <- module_tile_grid(scrna_cd8_all, pathway_list=setNames(gsea_subsets,gsea_subsets), subtypes=c('STIM-IFNG','STIM-GLYC','STIM-OXPHOS','STIM-UNTR','STIM-INACT','STIM-GNLY'), group.by='Subtype', dots=T)

plots$cd4_expanded_modules <- module_tile_grid(scrna_cd4_all, pathway_list=setNames(gsea_subsets,gsea_subsets), subtypes=c('STIM-IFNG','STIM-GLYC','STIM-OXPHOS','STIM-UNTR','STAT1-IRF1'), group.by='Subtype', dots=T)

# ALL subsets
#plots$cd8_module_tiles <- module_tile_grid(cd8_pruned, subtypes=cd8_order,
#    setNames(gsea_subsets,gsea_subsets), group.by='subtype_lbl', dots=F)
# plots$cd4_module_tiles <- module_tile_grid(cd4_pruned, subtypes=cd4_order,
#    setNames(gsea_subsets,gsea_subsets), group.by='subtype_lbl', dots=F)

```
## Supplemental per donor CAR plot

```{r}

cd8_barplot_data <- get_cluster_enrich_plot(
  cd8_pruned,
   'subtype_lbl', hide='K_only',
  facet_formula=formula(k_type~car), bar_position='dodge',
  by_vars=c('car','CD4v8','k_type','donor'))

cd4_barplot_data <- get_cluster_enrich_plot(
  cd4_pruned,
   'subtype_lbl', hide='K_only',
  facet_formula=formula(k_type~car), bar_position='dodge',
  by_vars=c('car','CD4v8','k_type','donor'))

donor_barplot <- function(data, subtype_order) {
  
  data[, mean_donor_frac := mean(frac), by=c('k_type','t_type','cluster','car')]
  data[, mean_rel_frac := log2(mean_donor_frac/mean(mean_donor_frac)), by=c('k_type','t_type','cluster')]
  
  max_log2_enrich <- log2(3)
  
  disp_data <- data[mean_frac > 0.01][, 
    car := factor(car, levels=umap_car_order)][,
     cluster := factor(cluster, levels=cluster_name_dt[subtype_order, on='Subtype']$subtype_lbl)]
  
  ggplot(disp_data) + 
      geom_bar(aes(
        x=cluster,
        y=mean_donor_frac,
        fill=mean_rel_frac),
      stat='identity',
      position='dodge',
      color=NA,
      width=1) +
    geom_point(aes(
      x=cluster,
      y=frac,
      color=donor)) + 
      facet_grid(formula(k_type~car), space='free', scales='free_y') + 
      labs(x='Enrichment vs mean car') +
      scale_y_continuous('Fraction of CAR in cluster',
        oob=scales::oob_squish, labels=label_percent(1)) +
      scale_fill_distiller('Log2 Enrichment\nvs mean CAR', palette='PRGn', direction=1,
        limits=c(-max_log2_enrich, max_log2_enrich), oob=scales::squish,
        breaks=c(-max_log2_enrich, 0, max_log2_enrich), labels=c('1/3x', '1x', '3x')) +
      theme_minimal() +
      coord_flip() +
      theme(
        panel.grid=element_blank(),
        axis.text.x = element_text(angle = 90, vjust = 0, hjust=1),
        panel.border=element_rect(color='grey70', fill=NA), 
        legend.key.width=unit(0.4, 'lines'))
}

plots$cd8_pct_barplots <- (
  (donor_barplot(cd8_barplot_data$data[k_type == 'cd19+'], rev(cd8_order)) + 
    theme(axis.title=element_blank())) / 
  (donor_barplot(cd8_barplot_data$data[k_type == 'baseline'], rev(cd8_order)) +
     theme(strip.text.x = element_blank()))
) + 
  plot_layout(heights=c(10, 4), guides='collect') & 
  theme(
    legend.position = 'bottom',
    legend.key.height=unit(0.4, 'lines'),
    legend.key.width=unit(2, 'lines'))

plots$cd4_pct_barplots <- (
  (donor_barplot(cd4_barplot_data$data[k_type == 'cd19+'], rev(cd4_order)) + 
    theme(axis.title=element_blank())) / 
  (donor_barplot(cd4_barplot_data$data[k_type == 'baseline'], rev(cd4_order)) +
     theme(strip.text.x = element_blank()))
) + 
  plot_layout(heights=c(8, 6), guides='collect') & 
  theme(
    legend.position = 'bottom',
    legend.key.height=unit(0.4, 'lines'),
    legend.key.width=unit(2, 'lines'))

```

### Innateness Plot

```{r}

innateness_dt <- fread(here::here('..','data','gutierrez_arcelus_et_al_2019-sup_dat_3.csv'))

innateness_sig_dt <- innateness_dt[
  CalledSignificant==1 &
  GENE_NAME %in% scrna_all@assays$SCT_INT@var.features & 
  GENE_NAME != 'AKR1C3']

innateness_scrna_dt <- data.table(FetchData(prune_cells(scrna_all),
    vars=c(names(scrna_all@meta.data), innateness_sig_dt$GENE_NAME),
    slot='scale.data'))

innateness_scrna_dt$innateness <- scale(
  as.matrix(innateness_scrna_dt[, innateness_sig_dt$GENE_NAME, with=F]) %*% 
    innateness_sig_dt$`InnatenessLevel_(Beta)`)

# add wnn umap
innateness_scrna_dt <- umap_dt[, list(cell, wnnUMAP_1, wnnUMAP_2)][innateness_scrna_dt, on='cell']

innateness_scrna_dt[, t_subtype := paste(t_type, Subtype, sep='.')]

innateness_boxplots <- (
  (ggplot(innateness_scrna_dt) +
    geom_boxplot(aes(x=Subtype, y=innateness, color=Subtype, fill=Subtype), alpha=0.5) + 
    facet_grid(t_type~., scale='free') +
    theme_minimal() +
    coord_flip() +
    scale_color_manual(values=all_colors) +
    scale_fill_manual(values=all_colors)) | 
  ggplot(innateness_scrna_dt) +
    geom_boxplot(aes(x=car, y=innateness, color=car, fill=car), alpha=0.5) + 
    facet_grid(t_type~., scale='free') +
    theme_minimal() +
    coord_flip() + 
    scale_color_manual(values=car_colors_umap) + 
    scale_fill_manual(values=car_colors_umap)
  ) & theme(legend.position='')

innateness_umap <- ggplot(
    innateness_scrna_dt[sample(.N)], aes(x=wnnUMAP_1, y=wnnUMAP_2, 
      color=innateness)) +
  rasterise(geom_point(size=inset_umap_point_size, alpha=0.5), dpi=300) + 
  scale_color_distiller('Lymphocyte Innateness\nZ-score\n(Gutierrez-Arcelus et al.)', palette='RdYlGn', direction=1, limits=c(-2,2), oob=scales::squish) +
  coord_fixed() +
  theme_minimal() + 
  guides(color = guide_colorbar(override.aes = list(size = 5, alpha=1), title.position='top', barheight = unit(0.4, 'line'))) +
  theme(
    legend.position=c(0,1),
    legend.direction = 'horizontal',
    legend.justification = c("left", "top"),
    axis.text=element_blank(),
    axis.line = element_blank(),
    axis.title=element_blank(),
    panel.grid=element_blank())

# innateness_minidot_plots <- (
#    / (
#   make_minidot_plot_car(
#     innateness_scrna_dt[sample(.N)], 'innateness', rescale_t_types = T, reorder_x=T) +
#     labs(y='CAR z-score') +
#     theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5))
# ) & theme(plot.margin= unit(c(0, 0, 0, 0), "points"))
# 
#  innateness_plot <- ((( & 
#       theme(legend.position='none', plot.margin= unit(c(0, 0, 0, 0), "points"))) /
#     innateness_minidot_plots) +
#       plot_layout(heights=c(2,1.3))
```

### Combine All Paper UMAP plot sets

```{r}

plots$paper_correlation_umap <- (
  interpolate_score(umap_dt, 500, 'Sheih_et_al_IRF1', pal_name='Greys') + 
    ggtitle('Sheih et al. 2020', 
      subtitle='Clonotypes enriched after\nengraftment in CAR patients') | 
  (interpolate_score(umap_dt, 500, 'Nicolet_et_al_CD29core1', pal_name='Greys') +
    ggtitle('Nicolet et al. 2020', 
      subtitle='CD29+ Cytotoxic signature,\npredicts survival in Melanoma')) |
  (interpolate_score(innateness_scrna_dt, 500, 'innateness', pal_name='Greys') + 
    ggtitle('Gutierrez-Arcelus et al. 2019', subtitle='Lymphocyte Innateness'))) &
  theme(
    legend.position='none',
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))

cluster_zscore_plots <- (((
  make_minidot_plot(umap_dt, 'Sheih_et_al_IRF1', rescale_t_types = T, reorder_x=T) + 
    labs(y='cluster z-score') + 
    theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5))) | 
  make_minidot_plot(umap_dt, 'Nicolet_et_al_CD29core1', rescale_t_types = T, reorder_x=T) |
  make_minidot_plot(innateness_scrna_dt[sample(.N)], 'innateness', 
    rescale_t_types = T, reorder_x=T)))

car_zscore_plots <- (((
  make_minidot_plot_car(umap_nocd30_dt, 'Sheih_et_al_IRF1', rescale_t_types = T, reorder_x=T) + 
    labs(y='CAR z-score') + 
    theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5))) | 
  make_minidot_plot_car(umap_nocd30_dt, 'Nicolet_et_al_CD29core1', rescale_t_types = T, reorder_x=T) |
  make_minidot_plot_car(innateness_scrna_dt[sample(.N)], 'innateness', 
    rescale_t_types = T, reorder_x=T)))

plots$combined_paper_umap <- (
  (plots$paper_correlation_umap & theme(plot.margin= unit(c(0, 0, 0, 0), "points"))) /
    cluster_zscore_plots / car_zscore_plots) +
      plot_layout(heights=c(4,1,1))

paper_umaps_vert <- ((
  interpolate_score(umap_dt, 500, 'Sheih_et_al_IRF1', pal_name='Greys') + 
    ggtitle('Clonotypes enriched after\nengraftment in CAR patients', subtitle='Sheih et al. 2020')) / 
  (interpolate_score(umap_dt, 500, 'Nicolet_et_al_CD29core1', pal_name='Greys') +
    ggtitle('CD29+ Cytotoxic signature,\npredicts survival in Melanoma', subtitle='Nicolet et al. 2020')) /
  (interpolate_score(innateness_scrna_dt, 500, 'innateness', pal_name='Greys') + 
    ggtitle('Lymphocyte Innateness', subtitle='Gutierrez-Arcelus et al. 2019'))) &
  theme(
    plot.title = element_text(size=rel(0.8)),
    plot.subtitle = element_text(size=rel(0.7)),
    legend.position='none',
    plot.margin = unit(c(5.5, 5.5, 0, 5.5), "points"))

combined_zscores <- function(datatable, var, datatable2=NULL) {
  if (is.null(datatable2)) datatable2 <- datatable
  (make_minidot_plot(datatable, var, rescale_t_types = T, reorder_x=T) + 
    theme(axis.title.y=element_text(angle = 90, vjust = 0.5, hjust=0.5)) +
      labs(y='z-score')) |  (
  make_minidot_plot_car(datatable2, var, rescale_t_types = T, reorder_x=T) + 
    theme(axis.title.y=element_blank())) 
}

paper_zscores_vert <- (combined_zscores(umap_dt, 'Sheih_et_al_IRF1', umap_nocd30_dt) / 
  combined_zscores(umap_dt, 'Nicolet_et_al_CD29core1', umap_nocd30_dt) / 
  combined_zscores(innateness_scrna_dt, 'innateness'))

plots$paper_umaps_combined_vert <- (paper_umaps_vert | paper_zscores_vert) + plot_layout(widths=c(1,2.3))



```

### Main Text Panel Assembly

```{r}

# fig_A <- plots$combined_umap
# #fig_B <- plots$combined_gene_heatmap
# fig_B <- plots$car_dotplot_horiz
# fig_D <- plots$combined_tile_plot
# fig_E <- plots$combined_paper_umap

plots$fig_assembled <- plot_grid(
  plot_grid(plots$combined_umap, 
    plot_grid(plots$car_dotplot, plots$combined_paper_umap, ncol=2, labels=c('C','D')),
    ncol=1, rel_heights=c(1,1.5), labels=c('A','')), 
  plot_grid(
    plots$combined_gene_heatmap,
    plots$combined_tile_plot+coord_fixed()+theme(
      legend.position='bottom',
      legend.key.height=unit(0.4, 'lines'),
      legend.key.width=unit(2, 'lines')), 
    ncol=1, rel_heights=c(1.7,1), labels=c('B','E')),
  ncol=2, rel_widths=c(2,1))

#width=1312&height=886&scale=1
plots$fig_assembled_2 <- plot_grid(
  plot_grid(plots$combined_umap, plots$car_dotplot_horiz,
    ncol=1, rel_heights=c(1.2,1), labels=c('A','C')), 
  plot_grid(plots$combined_paper_umap, plots$combined_tile_plot,
    ncol=1, rel_heights = c(1.2,1), labels=c('B','D')),
  rel_widths=c(2,1))

#width=1841&height=885&scale=1
plots$fig_assembled_3 <- plot_grid(
  plot_grid(plots$combined_umap, plots$car_dotplot_horiz,
    ncol=1, rel_heights=c(1.2,1), labels=c('A','B')), 
  plot_grid(plots$combined_paper_umap, plots$combined_tile_plot,
    ncol=1, rel_heights = c(1.2,1), labels=c('C','D')),
  plots$combined_gene_heatmap_boxed,
  ncol=3,
  labels=c('','','E'),
  rel_widths=c(1.6,1,1.2))

plots$fig_assembled_4 <- plot_grid(
  plot_grid(
    plot_grid(plots$combined_umap, plots$car_dotplot_horiz,
      ncol=1, rel_heights=c(1.3,1), labels=c('A','B')), 
    plot_grid(plots$combined_paper_umap, plots$combined_tile_plot,
      ncol=1, rel_heights = c(1.3,1), labels=c('C','D')),
    ncol=2,
    labels=c('',''),
    rel_widths=c(1.2,1)),
  plots$combined_gene_heatmap_boxed_horiz,
  nrow=2,
  labels=c('','E'),
  rel_heights=c(2,1))

plots$fig_assembled_5 <- plot_grid(
  plot_grid(
    plot_grid(plots$combined_umap_vert, plots$combined_paper_umap,
      ncol=2, rel_widths=c(1.3,1), labels=c('A','C')), 
    plot_grid(plots$car_dotplot_horiz, plots$combined_tile_plot,
      ncol=2, rel_widths = c(1.5,1), labels=c('B','D')),
    ncol=1,
    labels=c('',''),
    rel_heights=c(1.3,1)),
  plots$combined_gene_heatmap_boxed_horiz,
  nrow=2,
  labels=c('','E'),
  rel_heights=c(2,1))

#width=1280&height=1366&scale=1
plots$fig_assembled_6 <- plot_grid(
  plot_grid(
    #A
    ((plots$subtypes + theme(legend.position='none')) / 
      (plots$per_phase | plots$per_donor_sep | plots$activation2) + 
        plot_layout(widths=c(1,4,1))) + plot_layout(heights=c(5,1)),
    #C
    plots$car_dotplot_horiz,
    #D,E
    plot_grid(
      plots$citeseq_umaps,
      plots$combined_tile_plot,
      ncol=2,
      rel_widths=c(1.2,1),
      labels=c('D','E')
    ),
    ncol=1,
    rel_heights=c(1.2,1,0.8),
    labels=c('A','C','D')),
  plot_grid(
    plots$combined_gene_heatmap_boxed,
    plots$paper_umaps_combined_vert,
    ncol=1,
    rel_heights=c(1.2,0.8),
    labels=c('B','E')),
  rel_widths=c(1,1)
)

scaling_factor <- 1.6
ggsave(here::here(
  '..','figs','compiled','fig5.pdf'),
  plots$fig_assembled_6, device='pdf',
  width=8.5*scaling_factor,
  height=10*scaling_factor,
  units=c('in'),
  dpi=300)

```

## Other plots

```{r}

rna_by_subtype_and_car <- function(
    rna_name,
    separate_by_donor=F,
    slot='counts',
    assay='RNA',
    rescale_data=F) {
  orig_assay <- DefaultAssay(scrna_all)
  DefaultAssay(scrna_all) <- assay
  rna_data <- data.table(FetchData(prune_cells(scrna_all),
    vars=c(names(scrna_all@meta.data), rna_name), slot=slot))
  
  stopifnot(names(rna_data)[length(names(rna_data))] == rna_name)
  
  names(rna_data)[length(names(rna_data))] <- 'RNA_count'
  interact <- NULL
  if (separate_by_donor)
    interact <- 'donor'
  
  y_axis_name <- 'RNA expression'
  if (slot == 'counts') y_axis_name <- 'UMIs per cell'
  if (slot == 'scale.data') y_axis_name <- 'z-score'
  
  if (slot == 'scale.data' & rescale_data) {
    rna_data[, RNA_count := scale(RNA_count), by=c('k_type','donor')]
  }
    
  by_subset <- ggplot(rna_data,
      aes_string(x='Subtype', y='RNA_count', color='Subtype', shape=interact,
        group= paste0("interaction(", paste0(c(interact,'Subtype'), collapse = ", "), ")"))) + 
    stat_summary(fun = mean, geom = "point", position=position_dodge(0.5)) + 
    stat_summary(fun.data = mean_se, geom = "errorbar", position=position_dodge(0.5)) +
    facet_grid(t_type~., scale='free') +
    scale_color_manual(values=all_colors) +
    theme_minimal() +
    coord_flip() +
    labs(title=paste0(rna_name,' Expression by Subtype'), y=y_axis_name) +
    theme(
      panel.background=element_rect(fill=NA, color='grey50'),
      panel.grid=element_blank())
  
  by_car <- ggplot(rna_data, 
      aes_string(x='car', y='RNA_count', color='car', shape=interact,
        group= paste0("interaction(", paste0(c(interact,'car'), collapse = ", "), ")"))) +
    stat_summary(fun = mean, geom = "point", position=position_dodge(0.5)) + 
    stat_summary(fun.data = mean_se, geom = "errorbar", position=position_dodge(0.5)) +
    facet_grid(t_type~k_type) +
    scale_color_manual(values=car_colors_umap) +
    theme_minimal() +
    coord_flip() +
    labs(title=paste0(rna_name,' Expression by CAR'), y='UMIs per cell') +
    theme(
      panel.background=element_rect(fill=NA, color='grey50'),
      panel.grid=element_blank())
  
  DefaultAssay(scrna_all) <- orig_assay
  
  if (slot == 'scale.data') {
    by_car <- by_car + geom_hline(yintercept=0, linetype=3, color='grey30', size=0.6)
    by_subset <- by_subset + geom_hline(yintercept=0, linetype=3, color='grey30', size=0.6)
  }
  
  return((by_car + by_subset & theme(legend.position='none')) + plot_layout(widths=c(2,1)))
}

# BAFF-R Expression
rna_by_subtype_and_car('TNFRSF13C', separate_by_donor=T)
```
### Supp Panel Assembly

```{r}

supp_S5_top <- (
  plots$per_car_umap /
  plot_grid(
    plots$matched_clusters,
    plots$car_dotplot_unstim_vert,
    ncol=2, nrow=1, rel_widths=c(1,2.5))
  ) + plot_layout(heights=c(1.2,1))
scaling_factor <- 1.6

supp_S5_donors <- plot_grid(
  plots$cd4_pct_barplots & theme(legend.position='none', axis.title.x=element_blank()),
  plots$cd8_pct_barplots, ncol=1, nrow=2, rel_heights=c(1,1.3))

supp_S5 <- plot_grid(
  supp_S5_top,
  plot_grid(
    (supp_S5_donors | as_grob(plots$combined_adt_plot + theme(legend.position='bottom'))) + 
      plot_layout(widths=c(3,1))),
  ncol=1,
  rel_heights=c(1,1.2))

supp_S5 <- plot_grid(
  plots$per_car_umap,
  plot_grid(
    plot_grid(
      plot_grid(
        NULL,
        plots$car_dotplot_unstim + theme(plot.margin=margin(0,0,0,10, unit='pt')),
        NULL,
        plots$matched_clusters + 
          labs(fill='% Overlap of top 100 DE genes') +
          theme(
            legend.position='top',
            legend.key.height = unit(0.4, 'lines')),
        NULL,
        nrow=1,
        rel_widths=c(0.05,1.1,0.05,0.9,0.10)),
      supp_S5_donors,
      ncol=1,
      rel_heights=c(1.2,2)),
    plots$combined_adt_plot + 
      theme(
        legend.position='bottom',
        legend.key.height = unit(0.4, 'lines')),
    ncol=2,
    rel_widths=c(3,1.2)),
  plots$rna_umaps,
  nrow=3,
  rel_heights=c(1,3,0.8)
)

supp_S5_v2 <- plot_grid(
  
  #A
  plots$per_car_umap,
  
  #B/C
  plot_grid(
    NULL,
    plots$matched_clusters + 
      scale_fill_gradientn(
        '% overlap of\ntop 100\nDE genes',
        colors=c('white',brewer.pal(9,'Greens')),  label=label_percent()) +
      theme(
        legend.position='right',
        legend.key.width = unit(0.4, 'lines')),
    NULL,
    plots$car_dotplot_unstim + theme(plot.margin=margin(0,0,0,10, unit='pt')),
    NULL,
    nrow=1,
    rel_widths=c(0.05,1.25,0.25,1.5,0.25)),
    
  #D/E
  plot_grid(
    NULL,
    plots$combined_adt_plot + 
      theme(
        legend.position='bottom',
        legend.key.height = unit(0.4, 'lines')),
    NULL,
    supp_S5_donors,
    NULL,
    nrow=1,
    rel_widths=c(0.05,0.8,0.05,2,0.10)),
  
  #F
  plots$rna_umaps,
  nrow=4,
  rel_heights=c(1,1,1.9,0.8))

ggsave(here::here(
  '..','figs','compiled','figS5.pdf'),
  supp_S5_v2, device='pdf',
  width=8.5*scaling_factor,
  height=11*scaling_factor,
  units=c('in'),
  dpi=300)
```
